% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjoint_matriks.R
\name{adjoint_matriks}
\alias{adjoint_matriks}
\title{Pembentukan Adjoint Matriks}
\usage{
adjoint_matriks(x)
}
\arguments{
\item{x}{variabel matriks persegi}
}
\value{
Adjoint matriks
}
\description{
Fungsi ini akan melakukan pembentukan adjoint matriks.
}
\details{
Fungsi ini untuk melakukan pembentukan adjoint matriks, yaitu dengan membuat matriks kofaktor sesuai
dengan nilai dari kofaktor untuk baris dan kolom matriks tersebut kemudian dikalian dengan
(-1) pangkat penjumlahan dari baris dan kolom tesebut. Setelah mendapatkan
matriks kofaktor, selanjutnya melakukan transpose dari matriks kofaktor untuk mendapatkan
adjoint matriks.
}
\examples{
x <- matrix(c(0,3,2,1,-6,6,5,9,1),3,3)
adjoint_matriks(x)
}
