% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule_helpers.R
\name{advdate}
\alias{advdate}
\title{Advance Date Calculation}
\usage{
advdate(
  weekonemonday,
  week,
  topic = NULL,
  assignment = NULL,
  unit = "Module ",
  version = 1,
  ...
)
}
\arguments{
\item{weekonemonday}{The date of the first Monday of the schedule in "YYYY-MM-DD" format.}

\item{week}{The week number for which the dates need to be calculated.}

\item{topic}{A vector of topics for each week. Default is NULL.}

\item{assignment}{The specific assignment day, e.g., "Friday". Default is NULL.}

\item{unit}{The unit prefix for the week number. Default is "Module ".}

\item{version}{The version of the function to use (0 or 1). Default is 1. Allows for older version to be used.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A formatted string indicating the dates for the specified week and optional topic or assignment day.
}
\description{
This function calculates and formats dates for each day of a specified week, based on the start date of the first Monday.
}
\examples{
advdate("2023-01-09", 1)
advdate("2023-01-09", 2, topic = c("Intro", "Advanced Topics"))
advdate("2023-01-09", 3, assignment = "Friday")

}
