% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDetect1.R
\name{autoDetect1}
\alias{autoDetect1}
\title{Autodetection algorithm 1}
\usage{
autoDetect1(x, threshold = 2, val = "sst")
}
\arguments{
\item{x}{cleaned dataframe. See \code{\link{processOIData}()}.}

\item{threshold}{offshore-coast difference for upwelling}

\item{val}{what the SST column is called}
}
\value{
dataframe with upwelling TRUE/FALSE added
}
\description{
Take a data frame processed from data downloaded from ERDDAP and evaluates upwelling at a specific date.
}
\details{
This expects that only one date is in the dataframe. Run this code to
filter the dataframe by date.
\code{df_processed <- x \%>\% dplyr::filter(date == custom_date)}
}
\keyword{auto}
