% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDetect3.R
\name{autoDetect3}
\alias{autoDetect3}
\title{Autodetection algorithm 3}
\usage{
autoDetect3(
  x,
  threshold = 2,
  val = "sst",
  p = imageryML::sample_points$wintri$km100,
  smooth.method = "ksmooth",
  l.offshore = imageryML::buffer300$wintri$line,
  d.offshore = 50,
  l.coast = imageryML::buffer20$wintri$line,
  d.coast = 0,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{cleaned dataframe of SST values. See \code{\link{processOIData}()}. Note, this function might require that lon_0=0 in the crs.}

\item{threshold}{offshore-coast difference for upwelling}

\item{val}{what the SST column is called}

\item{p}{the coastal points as a \code{sp::SpatialPoints} object. Default is to use \code{km100} in \verb{data("sample_points)"} for the coastline where the islands are not removed.}

\item{smooth.method}{(optional) the smoothing to apply to the offshore line}

\item{l.offshore}{(optional) the offshore line as a \code{sp::SpatialLines} object.}

\item{d.offshore}{the radius of the circle to average around the offshore points.}

\item{l.coast}{(optional) the coast line as a \code{sp::SpatialLines} object.}

\item{d.coast}{the radius along the coast to average around the coast points. Operates only along the coast line.}

\item{na.rm}{whether to remove NAs in the raster when calculating the offshore SST. Used to deal with islands.}

\item{...}{extra smoothing parameters to pass to \code{\link{getNearestPointOnLine}()}.}
}
\value{
dataframe with upwelling TRUE/FALSE added
}
\description{
Take a data frame processed from data downloaded from ERDDAP and evaluates
upwelling by comparing the SST at a point along a smoothed coast to the average
SST at a point offshore. This uses a smoothed coastline, smoothed offshore line,
and coastal sample points that are part of the \strong{imageML} package. See \code{\link{smoothCoastData}}.
}
\details{
This expects that only one date is in the dataframe. Run this code to
filter the dataframe by date.
\code{df_processed <- x \%>\% dplyr::filter(date == custom_date)}

The coastal points and offshore line are saved data objects in \strong{imageryML}. See \code{vignette("Smooth Coastline Data")} for how these data were created.
}
\keyword{auto}
