\name{bovdata}
\alias{bovdata}
\title{
Little information about bovdata package!
}
\description{
The purpose of this package is to make available to the public the databases we built and now we use to create the portfolios. This data is composed by the balance sheets of publicly traded companies, their quotations and quotations adjusted by corporate events data.

This package is a product of the research project Construção de carteiras de variância mínima utilizando o modelo Fama-French (Construction of minimum variance portfolios using the Fama-French model), was firstly thought by Professor Regis Augusto Ely, who also idealized the format of the functions and their output, and it is being built by researcher Michel Ricardo Meyer.
}
\author{
Michel Ricardo Meyer - \email{michel.ricardo.meyer@gmail.com}

Regis Augusto Ely - \email{regisaely@gmail.com}
}
