% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer300_wintri.R
\docType{data}
\name{buffer300.wintri}
\alias{buffer300.wintri}
\title{300 km coastline buffer}
\format{
An object of class \code{list} of length 4.
}
\usage{
buffer300.wintri
}
\description{
A 300 km buffer in \code{sp::SpatialPolygons}, \code{sp::SpatialLines}, and
dataframe format in Winkel Tripel
projection.
}
\keyword{data}
