% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage2.R
\name{calculate_CI.ci_class}
\alias{calculate_CI.ci_class}
\title{Print a CI from an object of the class ci_class}
\usage{
calculate_CI.ci_class(x, conf = 0.95)
}
\arguments{
\item{x}{a vector containing the data from the sample.}

\item{conf}{Specification of the coverage percentage for the given operation}
}
\value{
the confidence intervals of sample mean
}
\description{
Print a Confidence interval from an object of the class ci_class which has been createdusing the make_ci_class function.
}
\details{
This function returns the confidence intervals of a sample mean and returns a named vector of length 2, where the first value is the lower_bound, the second value is the upper_bound.
}
\examples{
ci_example=make_ci_class(c(1,2,3,4,5))
calculate_CI.ci_class(ci_example)

}
