% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRFM.R
\name{calculateRFM}
\alias{calculateRFM}
\title{calcualteRFM}
\usage{
calculateRFM(data, weight_recency = 1, weight_frequency = 1,
  weight_monetary = 1)
}
\arguments{
\item{data}{- data}

\item{r}{- recency}
}
\value{
Returns a RFM model
}
\description{
Calcualte the weight of an RFM score: recency, frequency, monetary
}
\details{
\code{data} blablabla
}
