% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{change_height}
\alias{change_height}
\title{RJMCMC: Change height.}
\usage{
change_height(
  partition,
  heights,
  spikes,
  hyper.param,
  kappa,
  mu,
  kappa0,
  T.min = NULL,
  ISI.type = "Gamma",
  do.log,
  which.heights = "independent"
)
}
\arguments{
\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{spikes}{Data frame containing spike sequences.}

\item{hyper.param}{ISI parameter.}

\item{kappa}{parameter of priors height}

\item{mu}{parameter of priors height}

\item{kappa0}{parameter of priors height}

\item{T.min}{Refractory period.}

\item{ISI.type}{The ISI distribution.}

\item{do.log}{Flag, where if do.log is true the calculations are computed on the log scale.}

\item{which.heights}{Method for prior distribution of heights, either 'independent' or 'martingale'.}
}
\value{
New partition and heights
}
\description{
RJMCMC: Change height.
}
