% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_diagonalizable.R
\name{check_diagonalizable}
\alias{check_diagonalizable}
\title{Check Apakah Matriks dapat dilakukan Diagonalisasi}
\usage{
check_diagonalizable(A)
}
\arguments{
\item{A}{variable input matriks}
}
\description{
Fungsi ini akan menentukan apakah matriks dapat dilakukan diagonalisasi atau tidak.
}
\details{
Fungsi ini untuk menentukan apakah matriks dapat dilakukan diagonalisasi atau tidak.
}
\examples{
A = matrix(c(2,-1,0,-1,2,0,0,0,3),3,3)
check_diagonalizable(A)
}
