% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{check_feasible}
\alias{check_feasible}
\title{Check a step function is feasible.}
\usage{
check_feasible(
  d.spikes,
  partition = NA,
  heights = NA,
  x = NA,
  end.time = NA,
  T.min = NULL,
  max.T.min = NA
)
}
\arguments{
\item{d.spikes}{Data frame containing spike sequences}

\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{x}{The intensity function}

\item{end.time}{End of experiment time}

\item{T.min}{Refractory period, default set to NULL.}

\item{max.T.min}{If T.min != NULL, the maximum allowable refractory period.}
}
\description{
A function that checks that the partition and heights of a step function is feasible to be used as an intensity function. The function produces an error if the intensity is not feasible.
}
