% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{check_quietly}
\alias{check_quietly}
\title{Quietly Check Packages}
\usage{
check_quietly(...)
}
\arguments{
\item{...}{Arguments passed to purrr::quietly}
}
\value{
A function that when executed will return a list containing elements `result`, `output`, `warnings`, and `messages`.
}
\description{
Creates a wrapper around `devtools::check()` to execute package checking quietly.
This function captures and returns the output and errors without printing them directly to the console.
}
\examples{
\dontrun{
result <- check_quietly("myPackage")
}
}
\seealso{
`purrr::quietly()` for details on the structure of the returned list.
}
