% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkunits.R
\name{checkunits}
\alias{checkunits}
\title{Get or convert the units from a projection string}
\usage{
checkunits(x, units = "m")
}
\arguments{
\item{x}{a projection string}

\item{units}{the units you want to convert to}
}
\value{
list with crs with units added and units
}
\description{
Utility function to return the units.
}
\keyword{image}
