% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{conditioned_poisson}
\alias{conditioned_poisson}
\title{Poisson conditioned on max value.}
\usage{
conditioned_poisson(k, lambda, k.max, sum.of.prob)
}
\arguments{
\item{k}{k}

\item{lambda}{Parameter of the Poisson distribution.}

\item{k.max}{Maximum number of change points in RJMCMC.}

\item{sum.of.prob}{sum^k.max_i=0 Poi(k | lambda).}
}
\value{
probability of k
}
\description{
Poisson conditioned on max value.
}
