% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{conf_interval_pwc}
\alias{conf_interval_pwc}
\title{get the mean and 95\% conf interval for PWC method}
\usage{
conf_interval_pwc(
  partitions,
  heights,
  x,
  debug = FALSE,
  return.val = FALSE,
  method = "TRUE",
  steps = 1000,
  incProg = FALSE
)
}
\arguments{
\item{partitions}{za}

\item{heights}{za}

\item{x}{za}

\item{debug}{za}

\item{return.val}{za}

\item{method}{za}

\item{steps}{za}

\item{incProg}{za}
}
\value{
A list
}
\description{
get the mean and 95\% conf interval for PWC method
}
