% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraction_dilation.R
\name{contraction_dilatation}
\alias{contraction_dilatation}
\title{Mencari Kontraksi dan Dilatasi dari vektor R2 atau Vektor R3}
\usage{
contraction_dilatation(A, k)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi kontraksi dan dilatasi}

\item{k}{besaran skalar k}
}
\value{
Hasil transformasi kontraksi dan dilatasi
}
\description{
Fungsi ini akan melakukan pencarian kontraksi dan dilatasi dari vektor
berdasarkan besaran faktor skalar k.
}
\details{
Fungsi ini untuk melakukan pencarian kontraksi dan dilatasi dari vektor
berdasarkan besaran faktor skalar k. Jika besaran k berada di antara 0 dan 1
maka dilakukan transformasi kontraksi, sedangkan jika besaran k lebih besar
atau sama dengan 1 maka dilakukan transformasi dilatasi. Jika data yang dimasukan berjumlah 2 maka
dilakukan untuk vektor R2, jika berjumlah 3 maka dilakukan vektor R3.
}
\examples{
A = c(3,4)
k = 3
contraction_dilatation(A, k)
}
