% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{counting}
\alias{counting}
\title{Creates counting}
\usage{
counting(x, y)
}
\arguments{
\item{x}{numeric variable}

\item{y}{numeric variable}
}
\description{
`sum` returns the sum of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
counting(10, 3)
counting(20, 5)
}
