% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{cov.fn}
\alias{cov.fn}
\title{Square Exponential covariance function}
\usage{
cov.fn(t, sigma = sqrt(1000), l = 3)
}
\arguments{
\item{t}{Time difference}

\item{sigma}{Signal Variance}

\item{l}{Length Scale}
}
\value{
The Covariance
}
\description{
Square Exponential covariance function
}
