% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{covariance}
\alias{covariance}
\title{Calculate the covarience of two given points.}
\usage{
covariance(x, y, sigma_f2, sigma_v2, l)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{sigma_f2}{signal variance}

\item{sigma_v2}{noise variance}

\item{l}{length scale}
}
\value{
covariance
}
\description{
Calculate the covarience of two given points.
}
