% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_rule.R
\name{cramer_rule}
\alias{cramer_rule}
\title{Cramer Rule}
\usage{
cramer_rule(A, b)
}
\arguments{
\item{A}{variable matriks persegi}

\item{b}{variabel vektor}
}
\value{
Penyelesaian persamaan linear
}
\description{
Fungsi ini untuk menyelesaikan sistem persamaan linear dengan menggunakan metode Cramer's Rule.
}
\details{
Fungsi ini untuk melakukan penyelesaian sistem persamaan linear dengan menggunakan Cramer's Rule,
yaitu dengan cara mencari determinan terlebih dahulu untuk masing-masing A_i, yaitu matriks yang kolom ke-i
digantikan dengan nilai dari vektor y, kemudian untuk penyelesaian persamaan linear
dilakukan dengan membagi hasil determinan pada matriks A_i dengan hasil determinan matriks awal (A).
}
\examples{
A <- matrix(c(1,-3,-1,0,4,-2,2,6,3),3,3)
b <- c(6,30,8)
cramer_rule(A,b)
}
