% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CramerRule.R
\name{CramerRule}
\alias{CramerRule}
\title{Polynomial Interpolation}
\usage{
CramerRule(A, b)
}
\arguments{
\item{A}{squared matrix variable}

\item{b}{vector variable}
}
\description{
'Gauss Jordan' dengan secara bertahap
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
A <- matrix(c(1,-3,-1,0,4,-2,2,6,3),3,3)
b <- c(6,30,8)
CramerRule(A,b)
}
