% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{csv_to_spike}
\alias{csv_to_spike}
\title{Get spikes from .csv file of raw data.}
\usage{
csv_to_spike(
  filename,
  time.step = 2,
  transients = 2000,
  spike.occur = 0.6,
  closeness = 40,
  rm.trend = TRUE
)
}
\arguments{
\item{filename}{path to the csv file .}

\item{time.step}{Vector of time from the time series data, or if step size is regular the time step of the experiment.}

\item{transients}{transient periods where we want to remove spikes.}

\item{spike.occur}{threshold where we record spikes.}

\item{closeness}{Value which if two spikes are within this distance they are merged.}

\item{rm.trend}{Flag, where if TRUE we remove linear trends.}
}
\value{
data frame containing spike sequences as columns.
}
\description{
Get spikes from .csv file of raw data.
}
