% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{data_to_spike}
\alias{data_to_spike}
\title{Get spikes from raw data}
\usage{
data_to_spike(
  data,
  transients = 2000,
  time.step = 2,
  spike.occur = 0.6,
  val.change = NULL,
  reset = spike.occur,
  closeness = 40,
  rm.trend = TRUE
)
}
\arguments{
\item{data}{time series data.}

\item{transients}{transient periods where we want to remove spikes.}

\item{time.step}{Vector of time from the time series data, or if step size is regular the time step of the experiment.}

\item{spike.occur}{threshold where we record spikes.}

\item{val.change}{vector of times where the threshold value changes.}

\item{reset}{concentration at which we say a spike is over, default is set to spike.occur.}

\item{closeness}{Value which if two spikes are within this distance they are merged.}

\item{rm.trend}{Flag, where if TRUE we remove linear trends.}
}
\value{
spike sequence
}
\description{
Get spikes from raw data
}
