% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{death}
\alias{death}
\title{RJMCMC: Death of a change point}
\usage{
death(
  partition,
  heights,
  d.spikes,
  hyper.param,
  cvalue,
  kappa,
  mu,
  kappa0,
  k.max,
  sum.of.prob,
  lambda,
  T.min = NULL,
  max.T.min = NA,
  ISI.type = "Gamma",
  do.log,
  which.heights = "independent"
)
}
\arguments{
\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{d.spikes}{Data frame containing spike sequences.}

\item{hyper.param}{ISI parameter.}

\item{cvalue}{parameter c in the RJMCMC, to choose which event to perform in MCMC iteration.}

\item{kappa}{parameter of priors height}

\item{mu}{parameter of priors height}

\item{kappa0}{parameter of priors height}

\item{k.max}{Maximum number of change points in RJMCMC.}

\item{sum.of.prob}{sum^k.max_i=0 Poi(k | lambda).}

\item{lambda}{Parameter of the Poisson distribution.}

\item{T.min}{Refractory period.}

\item{max.T.min}{Maximal allowed T.min values from d.spikes.}

\item{ISI.type}{The ISI distribution.}

\item{do.log}{Flag, where if do.log is true the calculations are computed on the log scale.}

\item{which.heights}{Method for prior distribution of heights, either 'independent' or 'martingale'.}
}
\value{
new partition and heights
}
\description{
RJMCMC: Death of a change point
}
