% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{delta}
\alias{delta}
\title{Function for Dirac Delta function.}
\usage{
delta(x = 0, y = 0)
}
\arguments{
\item{x}{x}

\item{y}{y}
}
\value{
Dirac Delta function
}
\description{
Function for Dirac Delta function.
}
