% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_table.R
\name{desc_table}
\alias{desc_table}
\title{Descriptive statistics table}
\usage{
desc_table(data, group.var = NULL, cont.var = NULL, cat.var = NULL, paired = F)
group.var, cont.var and cat.var can be omitted in the inputs, as the function will ask for them
}
\arguments{
\item{data}{Dataframe to use}

\item{group.var}{String of the grouping variable column name. The grouping variable MUST have only 2 different values}

\item{cont.var}{Chr string (or vector) of the continuous variable/s column name to take into account}

\item{cat.var}{Chr string (or vector) of the categorical variable/s column name to take into account}

\item{paired}{Logical, are continuous variables paired?}
}
\value{
A table with the following columns: variable name - group_1 values (mean ± sd for continuous, n(\%) for categorical) - group_2 values (mean ± sd for continuous, n(\%) for categorical) - p-value - statistical method used
}
\description{
This function returns a descriptive statistics table for both categorical and continues variable for a 2-group dataset.
It calculates also the p-value choosing the right statistical analysis.
}
\examples{
desc_table(mtcars, group.var = "am", cont.var = c("mpg", "disp"), cat.var = c("cyl", "vs"), paired = F)

# cont.var and cat.var can be set as NA. Function will NOT ask for them
desc_table(mtcars, group.var = , cont.var = NA, cat.var = c("cyl", "vs"), paired = F)
}
\seealso{
\code{\link{desc_kable}} for descriptive kable
}
\author{
Matteo Miotto
}
