% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinan_cofactor.R
\name{determinan_cofactor}
\alias{determinan_cofactor}
\title{Determinan dengan ekspansi kofaktor}
\usage{
determinan_cofactor(x, n = 1, by_row = TRUE)
}
\arguments{
\item{x}{matriks persegi}

\item{n}{ekspansi kofaktor untuk elemen baris atau kolom ke berapa (default n = 1)}

\item{by_row}{apakah ekspansi kofaktor dilakukan secara ekspansi baris atau ekspansi kolom, jika TRUE maka dilakukan ekspansi baris, jika FALSE maka dilakukan ekspansi kolom (default by_row = TRUE)}
}
\value{
Nilai determinan dari matriks
}
\description{
Fungsi ini akan melakukan pencarian nilai determinan dari
suatu matriks persegi dengan menggunakan metode kofaktor dan disertai dengan langkah pengerjaannya.
}
\details{
Fungsi ini untuk melakukan pencarian nilai determinan dengan menggunakan
metode ekspansi kofaktor, Metode ekspansi kofaktor dilakukan dengan cara
melakukan ekspansi baris atau ekspansi kolom pada matriks.
}
\examples{
x <- matrix(c(0,3,2,1,-6,6,5,9,1),3,3)
determinan_cofactor(x,n=1,by_row=TRUE)
}
