% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Determinan_row_reduction.R
\name{Determinan_row_reduction}
\alias{Determinan_row_reduction}
\title{Determinan with row reduction}
\usage{
Determinan_row_reduction(x)
}
\arguments{
\item{x}{squared matrix variable}
}
\description{
'Determinan' dengan menggunakan metode row reduction secara bertahap
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
x <- matrix(c(0,3,2,1,-6,6,5,9,1),3,3)
Determinan_row_reduction(x)
}
