% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diiference_diviser.R
\name{diff_div}
\alias{diff_div}
\title{Calcul de la différence divisée pour l'interpolation de Newton}
\usage{
diff_div(x, y)
}
\arguments{
\item{x}{Un vecteur de valeurs x}

\item{y}{Un vecteur de valeurs y correspondant aux valeurs x}
}
\value{
La différence divisée des valeurs y
}
\description{
Calcul de la différence divisée pour l'interpolation de Newton
}
