% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_parellel_plane.R
\name{distance_parallel_plane}
\alias{distance_parallel_plane}
\title{Distance antara Bidang Paralel}
\usage{
distance_parallel_plane(plane1, plane2, by_first_plane = TRUE)
}
\arguments{
\item{plane1}{vektor dari bidang pertama}

\item{plane2}{vektor dari bidang kedua}

\item{by_first_plane}{penentuan bidang yang akan dijadikan nilai y dan z nya bernilai 0}
}
\value{
Nilai jarak dari 2 bidang paralel
}
\description{
Fungsi ini akan melakukan pencarian nilai jarak antara 2 bidang paralel.
}
\details{
Fungsi ini untuk melakukan pencarian pencarian nilai jarak antara 2 bidang paralel. by_first_plane digunakan
untuk menentukan bidang mana yang ingin dijadikan sebagai tujuannya. Jika bernilai TRUE
maka bidang pertama sebagai tujuan, sedangkan jika bernilai FALSE maka bidang
kedua akan dijadikan sebagai tujuan. Untuk vektor pada masing-masing bidang
harus memiliki ukuran yang sama.
}
\examples{
plane1 <- c(1,2,-2,-3)
plane2 <- c(2,4,-4,-7)
distance_parallel_plane(plane1, plane2)
}
