% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_point_line.R
\name{distance_point_line}
\alias{distance_point_line}
\title{Distance antara Titik dan Garis}
\usage{
distance_point_line(point, line)
}
\arguments{
\item{point}{vektor dari titik}

\item{line}{vektor dari garis}
}
\value{
Nilai jarak dari 1 titik ke 1 garis
}
\description{
Fungsi ini akan melakukan pencarian nilai jarak antara 1 titik dan 1 garis,
}
\details{
Fungsi ini untuk melakukan pencarian nilai jarak antara 1 titik dan 1 garis DImana titik
berada pada 2 dimensi (x,y) dan bidang dengan persamaan ax^2 + bx + c = 0 dituliskan
menjadi bentuk vektor (a,b,c)
}
\examples{
point <- c(-1,2)
line <- c(3,-4,9)
distance_point_line(point, line)
}
