% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{dividir}
\alias{dividir}
\title{Dividir um numero pelo outro}
\usage{
dividir(x, y)
}
\arguments{
\item{x}{um numero}

\item{y}{um numero}
}
\value{
a divisao de  \code{x} por \code{y}
}
\description{
\code{Dividir()} retorna a divisao de \code{x} e \code{y}.
}
\examples{
#dividir 6/2 = 3
dividir (6,2)

}
