% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_product.R
\name{dot_product}
\alias{dot_product}
\title{Dot Product Vektor}
\usage{
dot_product(u, v)
}
\arguments{
\item{u}{vektor input pertama}

\item{v}{vektor input pertama}
}
\value{
Plot dari 2 vektor
}
\description{
Fungsi ini akan melakukan pencarian nilai dot product vektor.
}
\details{
Fungsi ini untuk melakukan pencarian nilai dot product vektor, kemudian menghitung
nilai dari sudut antara 2 vektor dan akan menampilkan plot grafik dari 2 vektor yang dimasukkan.
}
\examples{
u <- c(3,0)
v <- c(4,4)
dot_product(u,v)
}
