% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qasi.R
\name{dqasi}
\alias{dqasi}
\title{Probability Density Function for Qasi Distribution}
\usage{
dqasi(t, beta, zeta, phi)
}
\arguments{
\item{t}{Numeric. The time or variable for which the PDF is to be computed.}

\item{beta}{Numeric. The scale parameter of the Qasi Distribution. Must be positive.}

\item{alpha}{Numeric. The shape parameter of the Qasi Distribution. Must be greater than 1.}

\item{lambda}{Numeric. The rate parameter of the Qasi Distribution. Must be positive.}
}
\value{
A numeric value representing the probability density for the given parameters at the specified \code{t}.
}
\description{
This function calculates the Probability Density Function (PDF) for the Qasi distribution, given the shape, scale, and rate parameters.
The Qasi distribution is a flexible probability model used in reliability analysis and survival studies, suitable for modeling real-world data involving failure times and lifetime distributions.
}
\details{
The probability density function describes the likelihood of a random variable taking a specific value. In the Qasi distribution, the PDF is useful for estimating failure rates and assessing reliability in various applications.
The formula for the Qasi distribution PDF is given by:
\deqn{f(t) = 4\alpha\theta^{2} t e^{-2\theta t }\bigg[1-\left(1+2\theta t \right)e^{-2\theta t }\bigg]^{\alpha-1}}
where \eqn{t} is the time or variable of interest.
}
\note{
Ensure that the parameters \code{alpha}, \code{beta}, and \code{lambda} are chosen appropriately to avoid numerical issues or undefined behavior.
}
\examples{
# Calculate the PDF for a given time point with specified parameters
pdf_value <- pdf_qasi(2, 1.5, 0.8, 0.3)

}
\references{
Ramzan, Q., Amin, M., & Faisal, M. (2022). Bayesian inference for modified Weibull distribution under simple step‐stress model based on type‐I censoring. \emph{Quality and Reliability Engineering International}, 38(2), 757-779.
}
\seealso{
\code{\link{dweibull}}, \code{\link{dexp}}
}
