% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{To plot the popularity of names in times}
\usage{
draw_a_name(Name, Sex)
}
\arguments{
\item{Name}{name from input}

\item{Sex}{sex from input}
}
\value{
the plot
}
\description{
To plot the popularity of names in times
}
