% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{To plot simultaneously the popularity of more than one name}
\usage{
draw_names(names)
}
\arguments{
\item{names}{names from input}
}
\value{
the plot
}
\description{
To plot simultaneously the popularity of more than one name
}
\examples{
draw_names(c("Vincent","Marie"))

}
