% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_2.R
\name{DRG_stat}
\alias{DRG_stat}
\title{Title}
\usage{
DRG_stat(stat)
}
\arguments{
\item{stat}{any statistics such as mean, median and sd}
}
\value{
values by drg code with the statistics used
}
\description{
Title
}
