% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total.R
\name{Ejt}
\alias{Ejt}
\title{R DATA ANALYSIS}
\usage{
Ejt(x)
}
\arguments{
\item{x}{is the function that will take the value of Ejt (main function).}
}
\value{
The solution for the specified exercise.
}
\description{
R DATA ANALYSIS
}
\examples{
# Retrieve the solution for exercise 3
Ejt(3)

# Retrieve the solution for exercise 6
Ejt(6)
}
