% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{exper.quantiles}
\alias{exper.quantiles}
\title{A function to get the quantiles given spikes and intensity function.}
\usage{
exper.quantiles(
  int.fn,
  end.time,
  spike.times,
  ISI.type,
  hyper.param,
  t.min = 0,
  do.log = FALSE
)
}
\arguments{
\item{int.fn}{int.fn}

\item{end.time}{end.time}

\item{spike.times}{spike.times}

\item{ISI.type}{ISI.type}

\item{hyper.param}{hyper.param}

\item{t.min}{t.min}

\item{do.log}{do.log}
}
\value{
experimental quantiles
}
\description{
A function to get the quantiles given spikes and intensity function.
}
