% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage.R
\name{fn_cos}
\alias{fn_cos}
\title{Print a Cosine}
\usage{
fn_cos(x, k)
}
\arguments{
\item{x}{real number}

\item{k}{degree of approximation}
}
\value{
An approximation of the Cosine value corresponding to a real number
}
\description{
Print an approximate Cosine value corresponding to a real number
}
\details{
This function returns an approximation of the Cosine value corresponding to a real number
}
\examples{
fn_cos(0,5)

}
