% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{forecast}
\usage{
forecast(locations, filepath = getwd(), download = T)
}
\arguments{
\item{locations}{SpatVector or data.frame. Locations for which forecast information shall be extracted.
If data.frame geographic coordinates must be provided in columns named "lon", "lat".
Optionally, the location names can be provided in a column "Name".}

\item{filepath}{Character. Enter path where downloaded files should be stored. Defaults to working directory.}

\item{download}{Logical. Indicates whether forecast data stored as .tif shall be downloaded. If TRUE, ca.
1.8 GB of data are downloaded. Defaults to TRUE.}
}
\value{
A list with a data table storing the forecast data, a plot allowing
checking these data visually, and a dashboard plot visualizing temperature
and precipitation.
}
\description{
implements a 5-day forecast based on data from the German Weather Service (DWD)
applicable for European locations. The forecast function allows download of forecast
data and creates a table and two plots checking these data as outputs. It
accesses the last_update() function which is also defined here.
}
\examples{
# Using a point SpatVector
locs    <- vect("C:/Docs/MIRO/GIS/Standorte/Standorte.shp")
fc      <- forecast(locations = locs, download = T)

# Using a data.frame
x <- data.frame(
lon = c(11:15),
lat = c(51:55)
)
fc      <- forecast(locations = x, download = F)

}
