% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage1.R
\name{fun_sin}
\alias{fun_sin}
\title{Print a Sine}
\usage{
fun_sin(x, k)
}
\arguments{
\item{x}{real number}

\item{k}{degree of approximation}
}
\value{
An approximation of the Sine value corresponding to a real number
}
\description{
Print an approximate Sine value corresponding to a real number
}
\details{
This function returns the Sine value corresponding to a real number
}
\examples{
fun_sin(3.1415/2,5)

}
