% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_jordan.R
\name{gauss_jordan}
\alias{gauss_jordan}
\title{Eliminasi Gauss-Jordan}
\usage{
gauss_jordan(A, b)
}
\arguments{
\item{A}{variabel matriks persegi}

\item{b}{variabel vektor}
}
\value{
Penyelesaian persamaan linear
}
\description{
Fungsi ini akan melakukan elimanasi Gauss-Jordan untuk menyelesaikan
sistem persamaan linear yang disertai dengan langkah pengerjaannya.
}
\details{
Fungsi ini untuk melakukan penyelesaian sistem persamaan linear dengan menggunakan eliminasi Gauss-Jordan.
Proses eliminasi Gauss-Jordan dilakukan dengan mengubah matriks koefisien menjadi matriks augmented,
kemudian melakukan operasi baris elementer hingga matriks tersebut menjadi matriks echelon baris,
lalu mengubah matriks tersebut menjadi matriks reduksi baris.
}
\examples{
A <- matrix(c(1,1,0,0,-1,2,1,0,1),3,3)
b <- c(4,-1,7)
gauss_jordan(A,b)
}
