% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_jordan_manual.R
\name{gauss_jordan_manual}
\alias{gauss_jordan_manual}
\title{Eliminasi Gauss-Jordan Secara Manual}
\usage{
gauss_jordan_manual(A, b)
}
\arguments{
\item{A}{variabel matriks persegi}

\item{b}{variabel vektor}
}
\value{
Penyelesaian persamaan linear
}
\description{
Fungsi ini akan melakukan elimanasi Gauss-Jordan untuk menyelesaikan sistem persamaan
linear secara manual dimana untuk prosesnya pengerjaannya ditentukan sendiri.
}
\details{
Fungsi ini untuk melakukan penyelesaian sistem persamaan linear dengan menggunakan eliminasi Gauss-Jordan.
Proses eliminasi Gauss-Jordan dilakukan sesuai dengan input fungsi yang diberikan, terdiri dari tukar() untuk
menukar 2 baris, kali() untuk mengalikan baris tertentu dengan nilai yang diberikan, lipat() untuk
menjumlahkan barsi, serta hasil() untuk melihat hasil akhir yang diberikan.
}
\examples{
A <- matrix(c(1,1,0,0,-1,2,1,0,1),3,3)
b <- c(4,-1,7)
tes = gauss_jordan_manual(A,b)
tes$lipat(1,2,-1)
tes$kali(2,-1)
tes$lipat(2,3,-2)
tes$kali(3,-1)
tes$lipat(3,2,-1)
tes$lipat(3,1,-1)
tes$hasil()
}
