% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussJordan.R
\name{GaussJordan}
\alias{GaussJordan}
\title{Gauss Jordan}
\usage{
GaussJordan(A, b)
}
\arguments{
\item{A}{squared matrix variable}

\item{b}{vector variable}
}
\description{
'Gauss Jordan' dengan secara bertahap
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
A <- matrix(c(1,1,0,0,-1,2,1,0,1),3,3)
b <- c(4,-1,7)
GaussJordan(A,b)
}
