\name{get.shares}
\alias{get.shares}
\title{
Internal Function!
}
\description{
This is a function used in other functions and must not work outside it's enviroments.
}
\usage{
get.shares(shares, envir = NULL)
}
\arguments{
  \item{shares}{
One or more shares
}
}
\value{
Returns a list with dataframes with the information of bovespa.
}
\seealso{
\code{\link{getPrices}}
\code{\link{getAdjPrices}}
}
\author{
Michel Ricardo Meyer \email{michel.ricardo.meyer@gmail.com}
}