% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{get.tau}
\alias{get.tau}
\title{A function that returns the ISI times under the transformation.}
\usage{
get.tau(R.t, spike.times, end.time)
}
\arguments{
\item{R.t}{R.t}

\item{spike.times}{spike.times}

\item{end.time}{end.time}
}
\value{
transformed ISI
}
\description{
A function that returns the ISI times under the transformation.
}
