% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{get_all_spikes_app}
\alias{get_all_spikes_app}
\title{Retrieve spikes from store and data.}
\usage{
get_all_spikes_app(data, store)
}
\arguments{
\item{data}{matrix containing raw time series data in columns, where the first column is time index.}

\item{store}{matrix containing threshold information}
}
\value{
matrix of spikes where each column is another spike sequence.
}
\description{
Retrieve spikes from store and data.
}
