% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{get_l}
\alias{get_l}
\title{get L from a filepath}
\usage{
get_l(filepath, save.file = T)
}
\arguments{
\item{filepath}{filepath}

\item{save.file}{save.file}
}
\value{
l values
}
\description{
get L from a filepath
}
