% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{get_QQ_KS}
\alias{get_QQ_KS}
\title{Get QQ and KS details}
\usage{
get_QQ_KS(filepath, model, return_val = T)
}
\arguments{
\item{filepath}{filepath}

\item{model}{model}

\item{return_val}{return_val}
}
\value{
details
}
\description{
Get QQ and KS details
}
