% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{get_QQ_KS_singleISI}
\alias{get_QQ_KS_singleISI}
\title{Return QQ and KS for single ISI.type and model type}
\usage{
get_QQ_KS_singleISI(filepath, model, ISI.type)
}
\arguments{
\item{filepath}{filepath}

\item{model}{model}

\item{ISI.type}{ISI.type}
}
\value{
QQ and KS
}
\description{
Return QQ and KS for single ISI.type and model type
}
