% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{get_slopes}
\alias{get_slopes}
\title{A function that returns the slopes of the individual spike cell data of QQ and KS plots.}
\usage{
get_slopes(data, spikes)
}
\arguments{
\item{data}{data}

\item{spikes}{spikes}
}
\value{
slopes
}
\description{
A function that returns the slopes of the individual spike cell data of QQ and KS plots.
}
