% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{get_spikes}
\alias{get_spikes}
\title{get_spikes()}
\usage{
get_spikes(d.spikes, i, rm.end.time = TRUE)
}
\arguments{
\item{d.spikes}{Data frame containing spike sequences.}

\item{i}{The index (column) denoting the required spike sequence.}

\item{rm.end.time}{A flag, where if true we remove the last spike from the spike sequence --- encase it corresponds to the experiment end time.}
}
\value{
A single spike sequence in a vector.
}
\description{
A function that returns a single spike sequence from a data frame of spikes.
}
