% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{get_spikes_app}
\alias{get_spikes_app}
\title{Get spikes for the shiny app.}
\usage{
get_spikes_app(
  data,
  time.step,
  thres,
  lin.trend,
  trans2_low,
  trans2_high,
  trans1,
  trans3,
  closeness
)
}
\arguments{
\item{data}{time series data.}

\item{time.step}{Vector of time from the time series data, or if step size is regular the time step of the experiment.}

\item{thres}{Threshold values in form c(spike.occur, val.change)}

\item{lin.trend}{Flag, where if TRUE we remove linear trends.}

\item{trans2_low}{lower bound of second transient period}

\item{trans2_high}{upper bound of secont transient period}

\item{trans1}{first transient period from 0 to trans1.}

\item{trans3}{lower bound of third transient period , where the upper bound is T, the end of experiment time.}

\item{closeness}{Value which if two spikes are within this distance they are merged.}
}
\value{
spikes
}
\description{
Get spikes for the shiny app.
}
