% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{get_X}
\alias{get_X}
\title{get_X()}
\usage{
get_X(partition, heights, d.spikes, T.min = NULL, rm.end.time = T)
}
\arguments{
\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{d.spikes}{Data frame containing spike sequences}

\item{T.min}{Refractory period, default set to NULL.}

\item{rm.end.time}{Flag, where if TRUE we remove the end time of the spike sequences.}
}
\value{
data frame containing the integral of the intensity between all spike times.
}
\description{
Calculates the integral of the intensity function, where the intensity is a piece-wise constant function.
}
