\name{getBalanceSheet}
\alias{getBalanceSheet}
\title{
Loads the Balance Sheets of the firms.
}
\description{
This functions load to R a data.frame with all balance sheets for the specified firms.
}
\details{
This function is an unfolding of \code{\link{getFinancialStatements}} and has the same author(s), arguments and details.
}
