% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{GetSigma}
\alias{GetSigma}
\title{Function that returns the covariance matrix Sigma.}
\usage{
GetSigma(tStop, l, sigma_f2, sigma_v2, steps)
}
\arguments{
\item{tStop}{end time}

\item{l}{length scale}

\item{sigma_f2}{signal variance}

\item{sigma_v2}{noise variance}

\item{steps}{steps}
}
\value{
Covariance matrix
}
\description{
Function that returns the covariance matrix Sigma.
}
