% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsAlongLine.R
\name{getStatsAlongLine}
\alias{getStatsAlongLine}
\title{Get the mean raster values along a line}
\usage{
getStatsAlongLine(r, p, l = NULL, d = 0, na.rm = TRUE)
}
\arguments{
\item{r}{the raster as a \code{raster::raster} object}

\item{p}{the points as a \code{sp::SpatialPoints} object}

\item{l}{the lines as a \code{sp::SpatialLines} object if \code{d} is not 0.}

\item{d}{distance to average along the line. Distance is in the units of the points and raster objects}

\item{na.rm}{whether to remove NAs in the raster when computing the average}
}
\value{
raster values around the points
}
\description{
This is applies \code{raster::extract()} but does some checking on the projection
of the inputs. Set \code{d=0} to get value at the points.
}
\details{
If there are multiple lines within the radius of \code{d}, then the average is over all lines covered
in that radius. The segments to average is determine by a circle around the points.
}
\seealso{
\code{\link{checkunits}()}
}
\keyword{auto}
