% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsAroundPoint.R
\name{getStatsAroundPoint}
\alias{getStatsAroundPoint}
\title{Get the mean raster values around a point}
\usage{
getStatsAroundPoint(p, r, d = 100, na.rm = FALSE)
}
\arguments{
\item{p}{the points as a \code{sp::SpatialPoints} object}

\item{r}{the raster as a \code{raster::raster} object}

\item{d}{distance to average around the point. Distance is in the units of the points and raster objects}

\item{na.rm}{whether to remove NAs in the raster when computing the average}
}
\value{
raster values around the points
}
\description{
This is applies \code{raster::extract()} but does some checking on the projection
of the inputs. Set \code{d=0} to get value at the points.
}
\seealso{
\code{\link{checkunits}()}
}
\keyword{auto}
