% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsAtPoint.R
\name{getValueAtPoint}
\alias{getValueAtPoint}
\title{Get the mean raster values around a point}
\usage{
getValueAtPoint(p, r, crs.to.use = crs.wintri)
}
\arguments{
\item{p}{the points as a \code{sp::SpatialPoints} object}

\item{r}{the raster as a \code{raster::raster} object}

\item{crs.to.use}{name of the projection. must be one that
the \strong{sp} package recognizes.}
}
\value{
raster value at the points
}
\description{
This is applies \code{raster::extract()} but does some checking on the projection
of the inputs.
}
\keyword{image}
