% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barras_porcent.R
\name{graf_bar}
\alias{graf_bar}
\title{Grafico De Barras Com Porcentagens}
\usage{
graf_bar(
  x,
  xlab = "",
  ylab = "",
  color = "Steelblue",
  ordenada = FALSE,
  size = 3.5,
  horizontal = FALSE
)
}
\arguments{
\item{x}{dados categoricos}

\item{xlab}{texto para o eixo x}

\item{ylab}{texto para o eixo y}

\item{color}{cor das barras do grafico}

\item{ordenada}{ordena as barras em forma descrescente}

\item{size}{comprimento}

\item{horizontal}{permite que as barras fiquem na horizontal}
}
\value{
grafico da variavel \code{x}
}
\description{
\code{graf_bar()} retorna um gráfico de barras.
}
\examples{
#grafico de barra simples graf_bar(dados)
graf_bar(ggplot2::diamonds$cut, xlab = "cut")

}
