% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newton_hermite.R
\name{hermite}
\alias{hermite}
\title{Fonction d'interpolation d'Hermite}
\usage{
hermite(x, y)
}
\arguments{
\item{x}{Un vecteur de points x.}

\item{y}{Un vecteur de points y.}
}
\value{
Une fonction qui évalue l'interpolant d'Hermite aux points donnés.
}
\description{
Fonction d'interpolation d'Hermite
}
\examples{
# Génération de 40 points entre -2*pi et 2*pi
x <- runif(min=-1,max=1,n=40)
#x <- seq(-2*pi, 2*pi, length.out = 400)
y <- exp(x)
# Utilisation de la fonction d'interpolation d'Hermite
h <- hermite(x, y)
# Génération de points supplémentaires pour le tracé
#y_plot <- sapply(x_plot, h)
# Création d'un dataframe pour ggplot
df_curve <- data.frame(x = x, y = y)
df_points <- data.frame(x = x, y = y)
# Création du graphique
p <- ggplot() +
  geom_line(data = df_curve, aes(x = x, y = y, color = "Interpolation d'Hermite")) +  # Ajout de la courbe
    geom_point(data = df_points, aes(x = x, y = y, color = "Points de données")) +  # Ajout des points de données
    labs(color = "Légende") +  # Ajout de la légende
    theme_minimal()  # Thème minimal pour un rendu propre
 # Affichage du graphique
 print(p)
}
