% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitung_vektor.R
\name{hitung_vektor}
\alias{hitung_vektor}
\title{Menghitung Operasi Dasar Vektor}
\usage{
hitung_vektor(u, v)
}
\arguments{
\item{u}{vektor input pertama}

\item{v}{vektor input kedua}
}
\value{
Nilai penjumlahan, pengurangan, dan perkalian dari vektor u dan vektor v.
}
\description{
Fungsi ini akan melakukan perhitungan operasi dasar dari 2 vektor berupa penjumlahan, pengurangan, dan perkalian.
}
\details{
Fungsi ini untuk menghitung penjumlahan, pengurangan dan perkalian dari 2 vektor yang diinput. Dimana dilakukan pengecekan terhadap jumlah data dari vektor yang dimasukkan.
}
\examples{
u <- c(1:4)
v <- c(5:8)
hitung_vektor(u,v)
}
