% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importdata.R
\name{import_erddap}
\alias{import_erddap}
\title{Download data from an ERDDAP server}
\usage{
import_erddap(
  start_date,
  stop_date,
  min_lat,
  max_lat,
  min_lon,
  max_lon,
  silent = FALSE
)
}
\arguments{
\item{start_date}{first date}

\item{stop_date}{final date}

\item{min_lat}{minimum latitude}

\item{max_lat}{maximum latitude}

\item{min_lon}{minimum longitude}

\item{max_lon}{maximum longitude}
}
\value{
The function saves a csv file and returns the data list invisibly.
}
\description{
Download data from CoastWatch ERDDAP server ncdcOisst21Agg data which is daily
optimal interpolation SST data.
}
\author{
Howard Baek
}
\keyword{data}
