% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{inc.fineness}
\alias{inc.fineness}
\title{A function to increase the fineness of the intensity function, by a factor of the user's choice.}
\usage{
inc.fineness(int.fn, factor)
}
\arguments{
\item{int.fn}{int.fn}

\item{factor}{factor}
}
\value{

}
\description{
A function to increase the fineness of the intensity function, by a factor of the user's choice.
}
