% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_mlverse}
\alias{install_mlverse}
\title{Install TensorFlow, Spark and MLflow}
\usage{
install_mlverse(version = c("auto", "gpu"))
}
\arguments{
\item{version}{The version to install. Currently \code{"auto"} or
\code{"gpu"}.}
}
\description{
Installs all 'mlverse' runtime dependencies.
}
