% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{install_quietly}
\alias{install_quietly}
\title{Quietly Install Packages}
\usage{
install_quietly(...)
}
\value{
A function that when executed will return a list containing elements `result`, `output`, `warnings`, and `messages`.
}
\description{
Creates a wrapper around `devtools::install()` to execute package installation quietly.
This function captures and returns the output and errors without printing them directly to the console.
}
\examples{
result <- install_quietly("myPackage")
}
\seealso{
`purrr::quietly()` for details on the structure of the returned list.
}
