% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolasi_polinomial.R
\name{interpolasi_polinomial}
\alias{interpolasi_polinomial}
\title{Interpolasi Polinomial}
\usage{
interpolasi_polinomial(x, y)
}
\arguments{
\item{x}{variabel vektor untuk koordinat x}

\item{y}{variabel vektor untuk koordinat y}
}
\value{
Persamaan interpolasi polinomial
}
\description{
Fungsi ini akan melakukan pencarian interpolasi dengan persamaan polinomial secara reduksi baris.
}
\details{
FUngsi ini untuk menemukan persamaan interpolasi polinomial dengan menggunakan reduksi baris
pada augmented matriks. Dimana untuk kolom pertama dari matriks yang dibuat berupa angka
1, kemudian untuk kolom lainnya diisikan sesuai dengan persamaan polinomial sebanyak jumlah
vektor x yang dimasukkan dikurangi 1 (n-1), dan untuk kolom terakhir akan digabungkan dengan
nilai pada vektor y.
}
\examples{
x <- c(1,2,3,4)
y <- c(3,-2,-5,0)
interpolasi_polinomial(x,y)
}
