% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invers_adjoint_matriks.R
\name{invers_adjoint_matriks}
\alias{invers_adjoint_matriks}
\title{Invers dengan adjoint Matriks}
\usage{
invers_adjoint_matriks(x)
}
\arguments{
\item{x}{variabel matriks persegi}
}
\value{
Invers dengan metode adjoint matriks
}
\description{
Fungsi ini akan melakukan pencarian nilai invers dengan menggunakan adjoint matriks.
}
\details{
Fungsi ini untuk melakukan pencarian nilai invers dengan menggunakan adjoint matriks, yaitu dengan membuat matriks kofaktor sesuai
dengan nilai dari kofaktor untuk baris dan kolom matriks tersebut kemudian dikalian dengan
(-1) pangkat penjumlahan dari baris dan kolom tesebut. Setelah mendapatkan
matriks kofaktor, selanjutnya melakukan transpose dari matriks kofaktor untuk mendapatkan
adjoint matriks. Setelah itu melakukan perkalian dari 1/determinan matriks dengan adjoint matriks.
}
\examples{
x <- matrix(c(0,3,2,1,-6,6,5,9,1),3,3)
invers_adjoint_matriks(x)
}
