% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invers_row_reduction.R
\name{invers_row_reduction}
\alias{invers_row_reduction}
\title{Invers dengan Row Reduction}
\usage{
invers_row_reduction(x)
}
\arguments{
\item{x}{matriks persegi}
}
\value{
Nilai invers dari matriks
}
\description{
Fungsi ini akan melakukan pencarian nilai invers dari suatu matriks persegi dengan menggunakan metode row reduction dan disertai dengan langkah pengerjaannya.
}
\details{
Fungsi ini untuk melakukan pencarian nilai invers dengan menggunakan metode row reduction, Metode
row reduction dilakukan dengan cara transformasi matriks menjadi matriks echelon baris dan kemudian
menerapkan subtitusi mundur untuk mendapatkan nilai invers.
}
\examples{
x <- matrix(1:4,2,2)
invers_row_reduction(x)
}
