% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{ISI_to_spike}
\alias{ISI_to_spike}
\title{Convert ISI times to spike times}
\usage{
ISI_to_spike(ISI)
}
\arguments{
\item{ISI}{ISI times.}
}
\value{
Spike times.
}
\description{
Convert ISI times to spike times
}
\examples{
ISI <- c(1,3,2,4,1,5)
ISI_to_spike(ISI)
}
