% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_iso.R
\name{iso_utc_str}
\alias{iso_utc_str}
\title{Format a POSIXct time as an ISO 8601 UTC timestamp string}
\usage{
iso_utc_str(x)
}
\arguments{
\item{x}{A POSIXct/POSIXt vector.}
}
\value{
A character vector of ISO 8601 UTC timestamps ending in `Z`.
}
\description{
Formats a POSIXct (or POSIXt) value as `"YYYY-MM-DDTHH:MM:SSZ"` using UTC.
}
\examples{
iso_utc_str(as.POSIXct("2026-01-12 21:00:00", tz = "UTC"))
}
