% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{itsa}
\alias{itsa}
\title{Rote learning.}
\usage{
itsa(n = runif(1, 1, 39), nota = FALSE)
}
\arguments{
\item{n}{number of repetitions}

\item{nota}{logical, to reinforce the statement}
}
\value{
nothing, invisibly
}
\description{
Repeat after me.
}
\examples{
itsa()
}

