% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinship.R
\name{kinship}
\alias{kinship}
\title{The function calculate additive matrix
major allele = 2, heterozygote = 1, minor allele = 0.}
\usage{
kinship(X)
}
\arguments{
\item{X}{A a set of SNP data.}
}
\description{
The function calculate additive matrix
major allele = 2, heterozygote = 1, minor allele = 0.
}
\examples{
data(DST2_maize)
K.t2 <- kinship(snp_DST2_maize)

}
