% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knnde.R
\name{knnde}
\alias{knnde}
\title{K-nearest neighbor density estimation using R}
\usage{
knnde(x, k, xrange, yrange)
}
\arguments{
\item{x}{data values which is two dimensional}

\item{k}{the order of the nearest neighbor}

\item{xrange, yrange}{the range of each dimension value}
}
\value{
estimated points' correspoding densities
}
\description{
the k-nearest neighbor density estimation for two-dimensional value using R, and this function needs R_package "FNN"
}
\examples{
\dontrun{
x <- as.matrix(faithful)
xrange <- seq(from = 1, to = 6, by = 0.1) 
yrange <- seq(from = 40, to = 100, by = 0.5)
k <- 5 
fit <- knnde(x, k, xrange, yrange)
}
}
