% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{likelihood_x}
\alias{likelihood_x}
\title{Likelihood of the intensity function.}
\usage{
likelihood_x(
  partition,
  heights,
  d.spikes,
  hyper.param,
  T.min = NULL,
  ISI.type = "Gamma",
  do.log = TRUE
)
}
\arguments{
\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{d.spikes}{Data frame containing spike sequences.}

\item{hyper.param}{Value of the ISI parameter.}

\item{T.min}{Refractory period, default set to NULL.}

\item{ISI.type}{The ISI distribution.}

\item{do.log}{Flag, where if do.log is true the calculations are computed on the log scale.}
}
\value{
Likelihood of the intensity function.
}
\description{
Likelihood of the intensity function.
}
