% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{log_pi_hyper_param}
\alias{log_pi_hyper_param}
\title{likelihood of the hyper parameter.}
\usage{
log_pi_hyper_param(
  value,
  partition,
  heights,
  d.spikes,
  prior.hyper.param,
  T.min = NULL,
  ISI.type = "Gamma"
)
}
\arguments{
\item{value}{ISI parameter.}

\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{d.spikes}{Data frame containing spike sequences.}

\item{prior.hyper.param}{Prior of the hyper parameter.}

\item{T.min}{Refractory period, default set to NULL.}

\item{ISI.type}{The ISI distribution.}
}
\value{
marginal of the hyper parameter
}
\description{
likelihood of the hyper parameter.
}
