% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{log_pi_l}
\alias{log_pi_l}
\title{Title}
\usage{
log_pi_l(l, x, end.time, sigma.f2, sigma.n2)
}
\arguments{
\item{l}{Length scale}

\item{x}{Intensity function}

\item{end.time}{Experiment end time}

\item{sigma.f2}{signal variance}

\item{sigma.n2}{noise variance}
}
\value{
Marginal of the length scale
}
\description{
Title
}
