% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{log_pi_Tmin}
\alias{log_pi_Tmin}
\title{Likelihood of the refractory period}
\usage{
log_pi_Tmin(
  d.spikes,
  T.min,
  max.T.min,
  hyper.param,
  partition = NA,
  heights = NA,
  x = NA,
  end.time = NA,
  prior.T.min = c(1, 0.01),
  ISI.type = "Gamma"
)
}
\arguments{
\item{d.spikes}{Data frame containing spike sequences}

\item{T.min}{Refractory period, default set to NULL.}

\item{max.T.min}{If T.min != NULL, the maximum allowable refractory period.}

\item{hyper.param}{Value of the ISI parameter.}

\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{x}{The intensity function}

\item{end.time}{End of experiment time}

\item{prior.T.min}{Prior for the refractory period.}

\item{ISI.type}{The ISI distribution.}
}
\value{
marginal of the refractory period
}
\description{
Likelihood of the refractory period
}
