% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{log_pi_x}
\alias{log_pi_x}
\title{Likelihood of the intensity function.}
\usage{
log_pi_x(
  d.spikes,
  hyper.param,
  partition = NA,
  heights = NA,
  x = NA,
  end.time = NA,
  T.min = NULL,
  max.T.min = NA,
  ISI.type = "Gamma",
  do.log = TRUE
)
}
\arguments{
\item{d.spikes}{Data frame containing spike sequences.}

\item{hyper.param}{Value of the ISI parameter.}

\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{x}{The intensity function}

\item{end.time}{End of experiment time}

\item{T.min}{Refractory period, default set to NULL.}

\item{max.T.min}{Maximal allowed T.min values from d.spikes.}

\item{ISI.type}{The ISI distribution.}

\item{do.log}{Flag, where if do.log is true the calculations are computed on the log scale.}
}
\value{
Likelihood of the intensity function.
}
\description{
Likelihood of the intensity function.
}
