% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage2.R
\name{make_ci_class}
\alias{make_ci_class}
\title{Make an object of the class "ci_class"}
\usage{
make_ci_class(x)
}
\arguments{
\item{x}{a vector}
}
\value{
an object of the class ci_class
}
\description{
This function returns an object of the class ci_class from a vector of data
}
\details{
This function returns an object of the class "ci_class"
}
\examples{
ci_example=make_ci_class(c(1,2,3,4,5))

}
