% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCoastBuffer.R
\name{makeCoastBuffer}
\alias{makeCoastBuffer}
\title{Make a coastline buffer}
\usage{
makeCoastBuffer(
  d = 300,
  units = "km",
  crs.to.use = "wintri",
  remove.holes = TRUE
)
}
\arguments{
\item{d}{distance from the coastline}

\item{units}{what units the distance is in}

\item{crs.to.use}{name of the projection. must be one that
the \strong{sp} package recognizes}

\item{remove.holes}{TRUE/FALSE. If TRUE (default), the interior
polygons are removed}
}
\value{
a list with a SpatialPolygons, SpatialLines and dataframe object
}
\description{
Make a line that is parallel to the world's coastlines at a specific distance. See \code{vignette("Smooth Coastline Data")} for how these data were created for the package.
}
\keyword{auto}
