% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{mcmc}
\alias{mcmc}
\title{RJMCMC: MCMC.}
\usage{
mcmc(
  spikes,
  end.time,
  iter,
  burn,
  k.max,
  lambda,
  kappa,
  mu,
  kappa0,
  hyper.param = c(1, 0.001),
  sigma.h = NULL,
  start.hyper = 1000,
  hyper.initial = 0.1,
  T.min.param = NULL,
  T.min.initial = 0.05,
  sigma.t = NULL,
  ISI.type = "Gamma",
  do.log = TRUE,
  show.iter = FALSE,
  which.heights = "independent"
)
}
\arguments{
\item{spikes}{Data frame containing spike sequences.}

\item{end.time}{End time of the experiment}

\item{iter}{Number of iterations to record}

\item{burn}{Number of iterations to burn before recording.}

\item{k.max}{Maximum number of change points in RJMCMC.}

\item{lambda}{Parameter of the Poisson distribution.}

\item{kappa}{parameter of priors height}

\item{mu}{parameter of priors height}

\item{kappa0}{parameter of priors height}

\item{hyper.param}{Value of the hyper parameter, or values for the gamma prior of the hyper parameter.}

\item{sigma.h}{Sigma used in RW-Metropolis for the ISI parameter.}

\item{start.hyper}{Iteration number to begin the ISI parameter.}

\item{hyper.initial}{Initial value of the ISI parameter.}

\item{T.min.param}{Value of the refractory period, or values for the gamma prior of the refractory period parameter.}

\item{T.min.initial}{Initial value of the refractory period.}

\item{sigma.t}{Sigma used in RW-Metropolis for the refractory period parameter.}

\item{ISI.type}{The ISI distribution.}

\item{do.log}{Flag, where if do.log is true the calculations are computed on the log scale.}

\item{show.iter}{Flag for whether to print the number of iterations complete into the console.}

\item{which.heights}{Method for prior distribution of heights, either 'independent' or 'martingale'.}
}
\value{
Iterations of the MCMC algorithm.
}
\description{
RJMCMC: MCMC.
}
