% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constant.R
\name{mcmc_constant}
\alias{mcmc_constant}
\title{MCMC algorithm when the intensity function is constant.}
\usage{
mcmc_constant(
  d.spikes,
  end.time,
  prior.hyper = c(1, 0.01),
  prior.x = c(1, 0.01),
  prior.tmin = c(1, 0.01),
  sigma.hyper,
  sigma.x,
  sigma.t.min,
  iter,
  burn,
  ISI.type,
  show.prog = T
)
}
\arguments{
\item{d.spikes}{d.spikes}

\item{end.time}{end.time}

\item{prior.hyper}{prior.hyper}

\item{prior.x}{prior.x}

\item{prior.tmin}{prior.tmin}

\item{sigma.hyper}{sigma.hyper}

\item{sigma.x}{sigma.x}

\item{sigma.t.min}{sigma.t.min}

\item{iter}{iter}

\item{burn}{burn}

\item{ISI.type}{ISI.type}

\item{show.prog}{show.prog}
}
\value{
mcmc
}
\description{
MCMC algorithm when the intensity function is constant.
}
\examples{
x <- 100
}
