% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{mcmc_gp}
\alias{mcmc_gp}
\title{Title za}
\usage{
mcmc_gp(
  d.spikes,
  end.time,
  iter,
  burn,
  steps,
  w,
  prior.x = c(1000, 1e-05),
  hyper.param = c(1, 0.001),
  ISI.type = "Gamma",
  sigma.h = NA,
  start.hyper = 1000,
  hyper.start = 1,
  l.param = 1,
  start.l = 1000,
  sigma.l = NA,
  initial.l = 1,
  T.min.param = NA,
  sigma.t = NA,
  x.start = NULL
)
}
\arguments{
\item{d.spikes}{za}

\item{end.time}{za}

\item{iter}{za}

\item{burn}{za}

\item{steps}{za}

\item{w}{za}

\item{prior.x}{za}

\item{hyper.param}{za}

\item{ISI.type}{za}

\item{sigma.h}{za}

\item{start.hyper}{za}

\item{hyper.start}{za}

\item{l.param}{za}

\item{start.l}{za}

\item{sigma.l}{za}

\item{initial.l}{za}

\item{T.min.param}{za}

\item{sigma.t}{za}

\item{x.start}{za}
}
\value{
za
}
\description{
Title za
}
