% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plot.R
\name{mean_plot}
\alias{mean_plot}
\title{Mean Plot}
\usage{
mean_plot(df, varx, vary)
}
\arguments{
\item{df}{a dataframe}

\item{varx}{a string name for variable x in the dataframe df}

\item{vary}{a string name for variable y in the datagrame df}
}
\value{
A scatter plot with \code{vary} versus \code{varx}
}
\description{
This function produces a scatter plot of \code{vary} versus \code{varx}
}
\examples{
df_test <- data.frame(x = rnorm(10),
                      y = rnorm(10))
mean_plot(df_test, 'x', 'y')

}
