% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{model.quantiles}
\alias{model.quantiles}
\title{A function that gets the qunatiles for the possion process of unit rate.}
\usage{
model.quantiles(N)
}
\arguments{
\item{N}{N}
}
\value{

}
\description{
A function that gets the qunatiles for the possion process of unit rate.
}
