% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{move_change_point}
\alias{move_change_point}
\title{RJMCMC: Move change point.}
\usage{
move_change_point(
  partition,
  heights,
  spikes,
  hyper.param,
  T.min = NULL,
  ISI.type = "Gamma",
  do.log
)
}
\arguments{
\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{spikes}{Data frame containing spike sequences.}

\item{hyper.param}{ISI parameter.}

\item{T.min}{Refractory period.}

\item{ISI.type}{The ISI distribution.}

\item{do.log}{Flag, where if do.log is true the calculations are computed on the log scale.}
}
\value{
new partition and heights
}
\description{
RJMCMC: Move change point.
}
