% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Motor Trend Car Road Tests}
\format{
A data frame with 32 observations on 11 (numeric) variables.
\describe{
\item{mpg}{	Miles/(US) gallon}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement (cu.in.)}
\item{hp}{Gross horsepower}
\item{drat}{Rear axle ratio}
\item{wt}{Weight (1000 lbs)}
}
}
\source{
\url{https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/mtcars}
}
\usage{
mtcars
}
\description{
The data was extracted from the 1974 Motor Trend
US magazine, and comprises fuel consumption and
10 aspects of automobile design and performance
for 32 automobiles (1973--74 models).
}
\note{
Henderson and Velleman (1981) comment in a footnote to Table 1: ‘Hocking original transcriber's noncrucial coding of the Mazda's
rotary engine as a straight six-cylinder engine and the Porsche's flat engine as a V engine, as well as the inclusion of the diesel
Mercedes 240D, have been retained to enable direct comparisons to be made with previous analyses.’
}
\keyword{datasets}
