% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Multiplicar de dois numeros}
\usage{
multiplicar(x, y)
}
\arguments{
\item{x}{um numero}

\item{y}{um numero}
}
\value{
a multiplica  \code{x} e \code{y}
}
\description{
\code{multiplicar()} retorna a multiplicacao de \code{x} e \code{y}.
}
\examples{
#subtrair 2*3 = 6
subtrair (2,3)

}
