% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiplicationbyStep.R
\name{MultiplicationbyStep}
\alias{MultiplicationbyStep}
\title{Multiplication Matrix with Step}
\usage{
MultiplicationbyStep(x, y)
}
\arguments{
\item{x}{numeric variable}

\item{y}{numeric variable}
}
\description{
`MultiplicationbyStep` returns the the step and resut of multiplication two matrix.
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
x <- matrix(1:4,2,2)
y <- matrix(5:8,2,2)
MultiplicationbyStep(x,y)
}
