% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatingSpikes.R
\name{MultiSpikes}
\alias{MultiSpikes}
\title{Simulate multiple spike sequences}
\usage{
MultiSpikes(
  end.time,
  int.fn,
  hyper,
  steps = 1000,
  T.min = NULL,
  ISI.type = "Gamma",
  multi = 10,
  add.end = TRUE,
  do.log = T
)
}
\arguments{
\item{end.time}{The length of time we want to simulate spikes}

\item{int.fn}{The intensity function}

\item{hyper}{The values of the ISI parameter}

\item{steps}{The number of steps to discretise time into}

\item{T.min}{The refractory period}

\item{ISI.type}{The ISI distribution}

\item{multi}{The number of spike sequences to generate}

\item{add.end}{Do we include the end time on the end of the spikes}

\item{do.log}{Do we do the calculations on the log scale}
}
\value{
Matrix containing multiple spikes sequences
}
\description{
Simulate multiple spike sequences
}
