% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap creates an interval using the following parameters.}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{<- number of iterations}

\item{x}{<- sample}

\item{fun}{<- function}

\item{alpha}{<- alpha}

\item{cx}{<- cex}
}
\value{
list
}
\description{
Bootstrap creates an interval using the following parameters.
}
