% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{mycltp}
\alias{mycltp}
\title{Make graphs of the Poisson distribution using certain parameters.}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{<- sample size}

\item{iter}{<- number of iterations}

\item{lambda}{}
}
\value{
graphs
}
\description{
Make graphs of the Poisson distribution using certain parameters.
}
