% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Creates maximum likelihood estimates.}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{<- sample vector}

\item{beta}{}
}
\value{
graphs
}
\description{
Creates maximum likelihood estimates.
}
