% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Display P-value areas}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{data variable}

\item{xmax}{maximum x value}

\item{n}{number of items taken from the sample}

\item{alpha}{the confidence interval percent}
}
\value{
a p-value and grpah
}
\description{
This will display the graph and p-value of a certain t0.
}
\examples{
set.seed(55);x1=rnorm(30,mean=25,sd=5)
tcalc=(mean(x1)-23)/(sd(x1)/sqrt(30))
mypvalue(tcalc,n=30,alpha=0.05)

}
