% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRead.R
\name{myread}
\alias{myread}
\title{Read in data and display in table.}
\usage{
myread(dird, csv)
}
\arguments{
\item{dird}{<- directory}

\item{csv}{<- string with .csv ending}
}
\value{
table
}
\description{
Read in data and display in table.
}
