% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{newton}
\alias{newton}
\title{Interpolation polynomiale par la méthode de Newton}
\usage{
newton(x, y)
}
\arguments{
\item{x}{Un vecteur de valeurs x}

\item{y}{Un vecteur de valeurs y correspondant aux valeurs x}

\item{newx}{Un vecteur de nouvelles valeurs x pour lesquelles nous voulons interpoler les valeurs y}
}
\value{
Un vecteur de nouvelles valeurs y interpolées pour les nouvelles valeurs x
}
\description{
Interpolation polynomiale par la méthode de Newton
}
\examples{
x <- c(1, 2, 3)
y <- c(1, 4, 9)
newx <- c(2.5)'
newy <- newton_interpolation(x, y, newx)'
print(newy)

}
