% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_vector.R
\name{norm_vector}
\alias{norm_vector}
\title{Norm atau panjang dari vektor}
\usage{
norm_vector(u)
}
\arguments{
\item{u}{variabel vektor}
}
\value{
Nilai panjang dari vektor
}
\description{
Fungsi ini akan melakukan pencarian nilai panjang dari suatu vektor
}
\details{
Fungsi ini untuk melakukan pencarian nilai panjang dari suatu vektor dengan
menjumlahkan nilai dari vektor yang sudah di pangkatkan 2, kemudian nilai dari
penjumlahan tersebut akan dikuadratkan.
}
\examples{
u <- c(1,2,1)
norm_vector(u)
}
