% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_holes.R
\name{only.holes}
\alias{only.holes}
\title{Keep only holes in polygons}
\usage{
only.holes(x)
}
\arguments{
\item{x}{A \code{sp::SpatialPolygons} object.}
}
\description{
Adapted from \strong{spatialEco} function \code{remove.holes()}
}
\keyword{image}
