% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{calculate CI}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{coef from glm}

\item{se}{se from glm}

\item{siglevel}{sig level}

\item{roundto}{decimal places}
}
\value{
or and ci
}
\description{
calculate CI
}
\examples{
coef = c(1:3)
se = c(0.1,0.2,0.3)
OR_95CI(coef = coef, se = se, siglevel =0.05,roundto=2)
}
