% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-package.R
\docType{package}
\name{package-package}
\alias{package}
\alias{package-package}
\title{package: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brainworkup/package}
  \item Report bugs at \url{https://github.com/brainworkup/package/issues}
}

}
\author{
\strong{Maintainer}: Joey Trampush \email{j.trampush@gmail.com} (\href{https://orcid.org/0000-0003-0830-508X}{ORCID}) [copyright holder]

}
\keyword{internal}
