% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-package.r, R/package.r
\docType{package}
\name{package}
\alias{package}
\alias{package-package}
\title{package.}
\usage{
package()
}
\value{
character string, the name of this package
}
\description{
package.

It's a package, not a library.
}
\details{
Heh.
}
\examples{
package()
}

