% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage.R
\docType{data}
\name{packageReadmeTemplate}
\alias{packageReadmeTemplate}
\title{This is package `package`.}
\format{An object of class \code{character} of length 1.}
\usage{
packageReadmeTemplate
}
\description{
This is package `package`.
}
\keyword{datasets}
