% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.2aov.R
\name{par.2aov}
\alias{par.2aov}
\title{Parametric/non parametric test evaluation for multivariate data}
\usage{
mrt.par.2aov(x, y, z, type.of.int = "+")
}
\arguments{
\item{x}{Numeric vector (or dataframe column)}

\item{y}{A vector of factors}

\item{z}{A vector of factors}

\item{tyoe.of.int}{"+" to compute the analysis on the two categorial variables only or "*" to compute also the analysis on the interaction}
}
\value{
\item{Shapiro categorie}{A dataframe containing W and p-value coming from shapiro.test of x in the two different categoric conditions}
\item{Bartlett}{A dataframe containing k-squared value, df and p-value coming from bartlett.test of x in the two different categoric conditions and, if type.of.int = "*", also from the interaction of the two categorical}
\item{Shapiro interazione}{optional, only if type.of.int = "*". Return the W and p-value coming from the shapiro test of all the possible interaction of y and z over x}
}
\description{
Get info about normality and variances of the given vectors in ordere to decide wether use 2-way ANOVA or Kruskal-Wallis
}
\examples{
par.2aov(rnorm(100), sample(rep(c(0,1), 50)), sample(rep(c("a", "b", "c", "d"), 25)), type.of.int = "*")
}
\author{
Matteo Miotto
}
