% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.lab.th.R
\name{pcr.lab.th}
\alias{pcr.lab.th}
\title{Create a PCR result file suitable for thesis lab}
\usage{
pcr.lab.th()
}
\value{
Excel file
}
\description{
This function returns an Excel file in which each sheet represent one gene, and in each sheet the table have the following columns: condition, ctmean, ± sign, ct SD
}
\author{
Matteo Miotto
}
