% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{PDF_tmin}
\alias{PDF_tmin}
\title{Calculate the probability density.}
\usage{
PDF_tmin(t, last.spike, hyper, end.time, x, X, ISI.type, t.min = 0, do.log = F)
}
\arguments{
\item{t}{The time of the current spike}

\item{last.spike}{The time of the previous spike.}

\item{hyper}{The ISI parameter value}

\item{end.time}{The end time of the experiment}

\item{x}{The intensity function --- defined in 0,end.time}

\item{X}{The integral of the intensity function}

\item{ISI.type}{The ISI distribution.}

\item{t.min}{t.min}

\item{do.log}{Flag for whether to calculate the pdf on the log scale.}
}
\value{
The probability density of a spike at time t given the last spike was at last.spike.
}
\description{
Calculate the probability density.
}
