% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{plot.95conf}
\alias{plot.95conf}
\title{Getmean + 95\% conf of intensity function}
\usage{
\method{plot}{`95conf`}(intensities, end.time, x = NULL, height = 0, return.val = FALSE, add = F)
}
\arguments{
\item{intensities}{intensities}

\item{end.time}{end.time}

\item{x}{x}

\item{height}{extra height on plot}

\item{return.val}{return.val}

\item{add}{add}
}
\value{
matrix
}
\description{
Getmean + 95\% conf of intensity function
}
