% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC integration.R
\name{plot.MCI}
\alias{plot.MCI}
\title{Plot for Monte-Carlo integration function}
\usage{
\method{plot}{MCI}(x)
}
\arguments{
\item{x}{A \code{list} of the class MCI}
}
\description{
Plot something
}
\examples{
obj = mc_int(x_range = c(0,1), fun = "x^2", B = 10^5)
plot(obj)
}
\author{
Stephane Guerrier
}
