% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{plot_everything}
\alias{plot_everything}
\title{Plot the important plots}
\usage{
plot_everything(filepath, model)
}
\arguments{
\item{filepath}{filepath}

\item{model}{model}
}
\value{
save plots
}
\description{
Plot the important plots
}
