% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolynomialInterpolation.R
\name{PolynomialInterpolation}
\alias{PolynomialInterpolation}
\title{Polynomial Interpolation}
\usage{
PolynomialInterpolation(x, y)
}
\arguments{
\item{x}{squared matrix variable}

\item{y}{vector variable}
}
\description{
'Polynomial Interpolation' dengan secara bertahap
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
x <- c(1,2,3,4)
y <- c(3,-2,-5,0)
PolynomialInterpolation(x,y)
}
