% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{pretty_install}
\alias{pretty_install}
\title{Pretty Install}
\usage{
pretty_install(...)
}
\arguments{
\item{...}{Arguments passed to devtools::install.}
}
\value{
Returns a cleaned list of output and messages.
}
\description{
Performs a quiet installation and cleans up the output, only displaying relevant information.
}
\examples{
install_results <- pretty_install("myPackage")
}
