% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage2.R
\name{print.ci_class}
\alias{print.ci_class}
\title{Print a ci_class object}
\usage{
\method{print}{ci_class}(x, ...)
}
\arguments{
\item{x}{belongs to ci_class}
}
\value{
the length of the underlying vector and the class type.
}
\description{
Print an object of the class ci_class
}
\details{
This function prints an object of the class ci_class, including the length of the underlying vector and the class type.
}
\examples{
ci_example=make_ci_class(c(1,2,3,4,5))
print.ci_class(ci_example)

}
