% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_function.R
\name{processing_function}
\alias{processing_function}
\title{processing_function}
\usage{
processing_function(filepath = "C:/Docs/MIRO/radiation_model/download_test")
}
\arguments{
\item{filepath}{Enter filepath where data to be processed is stored}
}
\description{
Processes downloaded data for given location and given time period.
}
