% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOIData.R
\name{processOIData}
\alias{processOIData}
\title{Process Optimal Interpolation data}
\usage{
processOIData(inputfile)
}
\arguments{
\item{inputfile}{a csv file like downloaded above}
}
\value{
return a processed data frame
}
\description{
Process a download from CoastWatch ERDDAP OI SST data. Mask out the Puget Sound
}
\details{
download.file("https://coastwatch.pfeg.noaa.gov/erddap/griddap/ncdcOisst21Agg.csv?sst\%5B(2019-01-01T12:00:00Z):1:(2020-01-01T12:00:00Z)\%5D\%5B(0.0):1:(0.0)\%5D\%5B(35.125):1:(52.125)\%5D\%5B(228.125):1:(238.125)\%5D", "data/pilot_data.csv")
}
\keyword{upwelling}
