% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_axis.R
\name{projection_axis}
\alias{projection_axis}
\title{Mencari Proyeksi dari vektor R2}
\usage{
projection_axis(A, axis)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi proyeksi}

\item{axis}{sumbu proyeksi terhadap vektor yang akan dilakukan}
}
\value{
Hasil transformasi proyeksi
}
\description{
Fungsi ini akan melakukan pencarian proyeksi dari vektor
berdasarkan pada axis.
}
\details{
Fungsi ini untuk melakukan pencarian proyeksi dari vektor
berdasarkan pada axis. Jika axis bernilai 'x', maka
akan dilakukan proyeksi terhadap sumbu x, jika axis bernilai 'y',
maka akan dilakukan proyeksi terhadap sumbu y.
}
\examples{
A = c(3,4)
axis = "x"
projection_axis(A, axis)
}
