% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_plane.R
\name{projection_plane}
\alias{projection_plane}
\title{Mencari Proyeksi dari vektor R3}
\usage{
projection_plane(A, plane)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi proyeksi}

\item{plane}{bidang proyeksi terhadap vektor yang akan dilakukan}
}
\value{
Hasil transformasi proyeksi
}
\description{
Fungsi ini akan melakukan pencarian proyeksi dari vektor
berdasarkan pada plane.
}
\details{
Fungsi ini untuk melakukan pencarian proyeksi dari vektor
berdasarkan pada plane. Jika plane bernilai 'xz', maka
akan dilakukan proyeksi terhadap bidang xz, jika plane bernilai 'xz',
maka akan dilakukan proyeksi terhadap bidang xz, sedangkan jika plane bernilai
'yz' maka akan dilakukan proyeksi terhadap bidang yz.
}
\examples{
A = c(3,4,5)
plane = "xz"
projection_plane(A, plane)
}
