% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_vector.R
\name{projection_vector}
\alias{projection_vector}
\title{Mencari Proyeksi dari vektor R2 dan R3}
\usage{
projection_vector(A, to)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi proyeksi}

\item{to}{proyeksi terhadap vektor yang akan dilakukan}
}
\value{
Plot hasil transformasi proyeksi
}
\description{
Fungsi ini akan melakukan pencarian proyeksi dari vektor
berdasarkan pada axis atau plane.
}
\details{
Fungsi ini untuk melakukan pencarian proyeksi dari vektor
berdasarkan pada axis atau plane.Jika axis bernilai 'x', maka
akan dilakukan proyeksi terhadap sumbu x, jika axis bernilai 'y',
maka akan dilakukan proyeksi terhadap sumbu y. Jika plane bernilai 'xz', maka
akan dilakukan proyeksi terhadap bidang xz, jika plane bernilai 'xz',
maka akan dilakukan proyeksi terhadap bidang xz, sedangkan jika plane bernilai
'yz' maka akan dilakukan proyeksi terhadap bidang yz.
}
\examples{
A = c(3,4)
to = "x"
projection_vector(A, to)

A = c(3,4,5)
to = "xz"
projection_vector(A, to)

}
