% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proyeksi_vektor.R
\name{proyeksi_vektor}
\alias{proyeksi_vektor}
\title{Proyeksi Vektor}
\usage{
proyeksi_vektor(u, a)
}
\arguments{
\item{u}{vektor input pertama}

\item{a}{vektor input pertama}
}
\value{
Nilai proyeksi skalar dan proyeksi ortogonal.
}
\description{
Fungsi ini akan melakukan pencarian nilai proyeksi skalar dan proyeksi ortogonal.
}
\details{
Fungsi ini untuk melakukan pencarian pencarian nilai proyeksi skalar dan proyeksi ortogonal,
dimana untuk dua vektor yang tidak berhimpit sehingga salah satu vektor bisa
diproyeksikan ke vektor lainnya.
}
\examples{
u <- c(0,0,1)
a <- c(0,2,2)
proyeksi_vektor(u,a)
}
