% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{q.t}
\alias{q.t}
\title{A function to calclate q(t), the conditional intensity function.}
\usage{
q.t(int.fn, end.time, spike.times, ISI.type, hyper, t.min = 0, do.log = TRUE)
}
\arguments{
\item{int.fn}{int.fn}

\item{end.time}{end.time}

\item{spike.times}{spike.times}

\item{ISI.type}{ISI.type}

\item{hyper}{hyper.param}

\item{t.min}{t.min}

\item{do.log}{do.log}
}
\value{
q(t)
}
\description{
A function to calclate q(t), the conditional intensity function.
}
\examples{
100
}
