% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{QQ_KS_data}
\alias{QQ_KS_data}
\title{Get the QQ and KS quantiles}
\usage{
QQ_KS_data(
  spikes.df,
  int.fn,
  end.time,
  ISI.type,
  hyper.param,
  t.min = 0,
  rm.last.spike = TRUE,
  do.log = TRUE
)
}
\arguments{
\item{spikes.df}{spikes}

\item{int.fn}{int.fn}

\item{end.time}{end.time}

\item{ISI.type}{ISI.type}

\item{hyper.param}{hyper.param}

\item{t.min}{t.min}

\item{rm.last.spike}{rm.last.spike}

\item{do.log}{do.log}
}
\value{
quantiles
}
\description{
Get the QQ and KS quantiles
}
