% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQ_KS.R
\name{R.t}
\alias{R.t}
\title{A function that returns R(t), the integrated intensity function.}
\usage{
R.t(q, time.step)
}
\arguments{
\item{q}{q}

\item{time.step}{time.step}
}
\value{

}
\description{
A function that returns R(t), the integrated intensity function.
}
