% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{raw_spike_seq}
\alias{raw_spike_seq}
\title{Convert time series to spikes}
\usage{
raw_spike_seq(
  data,
  time.step,
  spike.occur,
  val.change = NULL,
  reset = spike.occur
)
}
\arguments{
\item{data}{data}

\item{time.step}{time.step}

\item{spike.occur}{spike.occur}

\item{val.change}{val change}

\item{reset}{reset}
}
\value{
spike times
}
\description{
Convert time series to spikes
}
