% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_multi.R
\name{read_excel_multi}
\alias{read_excel_multi}
\title{Import all sheet from an excel file}
\usage{
read_excel_multi(file)
}
\arguments{
\item{file}{path to xlsx file}
}
\value{
a list
}
\description{
Import all sheet from an excel file
}
\examples{
\dontrun{
read_excel_multi("/folder/toxlsxfile")
}
}
