% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection_axis.R
\name{reflection_axis}
\alias{reflection_axis}
\title{Mencari Refleksi dari vektor R2}
\usage{
reflection_axis(A, axis)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi refleksi}

\item{axis}{sumbu refleksi terhadap vektor yang akan dilakukan}
}
\value{
Hasil transformasi refleksi
}
\description{
Fungsi ini akan melakukan pencarian refleksi dari vektor
berdasarkan pada axis ataupun garis.
}
\details{
Fungsi ini untuk melakukan pencarian refleksi dari vektor
berdasarkan pada axis ataupun garis. Jika axis bernilai 'x', maka
akan dilakukan refleksi terhadap sumbu x, jika axis bernilai 'y',
maka akan dilakukan refleksi terhadap sumbu y, sedangkan jika axis bernilai
'line' maka akan dilakukan refleksi terhadap line y = x.
}
\examples{
A = c(3,4)
axis = "x"
reflection_axis(A, axis)
}
