% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection_plane.R
\name{reflection_plane}
\alias{reflection_plane}
\title{Mencari Refleksi dari vektor R3}
\usage{
reflection_plane(A, plane)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi refleksi}

\item{plane}{bidang refleksi terhadap vektor yang akan dilakukan}
}
\value{
Hasil transformasi refleksi
}
\description{
Fungsi ini akan melakukan pencarian refleksi dari vektor
berdasarkan pada plane.
}
\details{
Fungsi ini untuk melakukan pencarian refleksi dari vektor
berdasarkan pada plane. Jika plane bernilai 'xz', maka
akan dilakukan refleksi terhadap bidang xz, jika plane bernilai 'xz',
maka akan dilakukan refleksi terhadap bidang xz, sedangkan jika plane bernilai
'yz' maka akan dilakukan refleksi terhadap bidang yz.
}
\examples{
A = c(3,4,5)
plane = "xz"
reflection_plane(A, plane)
}
