% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflection_vector.R
\name{reflection_vector}
\alias{reflection_vector}
\title{Mencari Refleksi dari vektor R2 dan R3}
\usage{
reflection_vector(A, to)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi refleksi}

\item{to}{refleksi terhadap vektor yang akan dilakukan}
}
\value{
Hasil transformasi refleksi
}
\description{
Fungsi ini akan melakukan pencarian refleksi dari vektor
berdasarkan pada axis, garis, ataupun plane.
}
\details{
Fungsi ini untuk melakukan pencarian refleksi dari vektor
berdasarkan pada axis, garis, ataupun plane. Jika axis bernilai 'x', maka
akan dilakukan refleksi terhadap sumbu x, jika axis bernilai 'y',
maka akan dilakukan refleksi terhadap sumbu y, sedangkan jika axis bernilai
'line' maka akan dilakukan refleksi terhadap line y = x. Jika plane bernilai 'xz', maka
akan dilakukan refleksi terhadap bidang xz, jika plane bernilai 'xz',
maka akan dilakukan refleksi terhadap bidang xz, sedangkan jika plane bernilai
'yz' maka akan dilakukan refleksi terhadap bidang yz.
}
\examples{
A = c(3,4)
to = "x"
reflection_vector(A, to)

A = c(3,4,5)
to = "xz"
reflection_vector(A, to)
}
