% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{remove_transients}
\alias{remove_transients}
\title{Remove transient periods}
\usage{
remove_transients(spikes, time)
}
\arguments{
\item{spikes}{spikes with end.time adjoined to the end}

\item{time}{the region/s of transients}
}
\value{
spikes with transients removed
}
\description{
Remove transient periods
}
\examples{
spikes <- c(0.1,0.3,0.8,0.9,2,4,8,15,18,23)

remove_transients(spikes, 1) # remove transients at start
remove_transients(spikes,c(0,7.5,16)) # remove transients in middle
remove_transients(spikes, c(0,16,25)) # remove transients at end
}
