% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{remove_trend}
\alias{remove_trend}
\title{Remove Linear trends}
\usage{
remove_trend(ISI, rm.end.time = TRUE)
}
\arguments{
\item{ISI}{The ISI times}

\item{rm.end.time}{Flag for whether we need to remove the last entry (end.time) of the ISI times.}
}
\value{
ISI.times with linear trend removed.
}
\description{
Remove Linear trends
}
\examples{
ISI <- c(1,3,2,4,1,5)
ISI.no <- 1:6
remove_trend(ISI, rm.end.time=FALSE)
stats::lm(ISI.no ~ remove_trend(ISI, rm.end.time=FALSE)) # Show the linear trend has been removed.
}
