% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_clockwise.R
\name{rotation_clockwise}
\alias{rotation_clockwise}
\title{Mencari Rotasi dari vektor R2}
\usage{
rotation_clockwise(A, alfa)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi rotasi}

\item{alfa}{besaran sudut}
}
\value{
Hasil transformasi rotasi
}
\description{
Fungsi ini akan melakukan pencarian rotasi dari vektor
berdasarkan searah jarum jam dan besaran sudut.
}
\details{
Fungsi ini untuk melakukan pencarian rotasi dari vektor
berdasarkan searah jarum jam dan besaran sudut.
}
\examples{
A = c(3,4)
alfa = 60
rotation_clockwise(A, alfa = alfa )
}
