% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_counterclockwise.R
\name{rotation_counterclockwise}
\alias{rotation_counterclockwise}
\title{Mencari Rotasi dari vektor R3}
\usage{
rotation_counterclockwise(A, alfa, axis)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi rotasi}

\item{alfa}{besaran sudut}

\item{axis}{sumbu rotasi terhadap vektor yang akan dilakukan}
}
\value{
Hasil transformasi rotasi
}
\description{
Fungsi ini akan melakukan pencarian rotasi dari vektor
berdasarkan berlawanan arah jarum jam, axis, dan besaran sudut.
}
\details{
Fungsi ini untuk melakukan pencarian rotasi dari vektor
berdasarkan berlawanan arah jarum jam, axis, dan besaran sudut. Jika axis bernilai 'x', maka
akan dilakukan rotasi terhadap sumbu x, jika axis bernilai 'y',
maka akan dilakukan rotasi terhadap sumbu y, dan jika axis bernilai 'z', maka
akan dilakukan rotasi terhadap sumbu 'z'.
}
\examples{
A = c(3,4,5)
alfa = 60
axis = "x"
rotation_counterclockwise(A, alfa = alfa, axis = axis )
}
