% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_vector.R
\name{rotation_vector}
\alias{rotation_vector}
\title{Mencari Rotasi dari vektor R2 dan R3}
\usage{
rotation_vector(A, alfa, axis = NULL)
}
\arguments{
\item{A}{vektor input yang ingin dilakukan transformasi rotasi}

\item{alfa}{besaran sudut}

\item{axis}{sumbu rotasi terhadap vektor yang akan dilakukan}
}
\value{
Plot hasil transformasi rotasi
}
\description{
Fungsi ini akan melakukan pencarian rotasi dari vektor
berdasarkan berlawanan arah jarum jam, axis, dan besaran sudut pada dua dimensi atau tiga dimensi.
}
\details{
Fungsi ini untuk melakukan pencarian rotasi dari vektor
berdasarkan berlawanan arah jarum jam, axis, dan besaran sudut pada dua dimensi atau tiga dimensi.
Jika berada pada tiga dimensi maka ditambahkan dengan axis atau arah rotasi
terhadap sumbu. axis bernilai 'x', maka
akan dilakukan rotasi terhadap sumbu x, jika axis bernilai 'y',
maka akan dilakukan rotasi terhadap sumbu y, dan jika axis bernilai 'z', maka
akan dilakukan rotasi terhadap sumbu 'z'.
}
\examples{
A = c(3,4)
alfa = 60
rotation_vector(A, alfa = alfa )


A = c(3,4,5)
alfa = 60
axis = "x"
rotation_vector(A, alfa = alfa, axis = axis )
}
