% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qasi.R
\name{rqasi}
\alias{rqasi}
\title{Quantile Function for Modified Weibull Distribution (MWD)}
\usage{
rqasi(n, alpha, beta, lambda)
}
\arguments{
\item{n}{Integer. The number of random observations to generate.}

\item{alpha}{Numeric. The shape parameter of the Modified Weibull Distribution (MWD). Must be greater than 1.}

\item{beta}{Numeric. The scale parameter of the Modified Weibull Distribution (MWD). Must be positive.}

\item{lambda}{Numeric. The rate parameter of the Modified Weibull Distribution (MWD). Must be positive.}
}
\value{
A numeric vector of length \code{n} containing generated observations from the Modified Weibull Distribution (MWD).
}
\description{
This function generates random observations by calculating the Quantile Function for the Modified Weibull Distribution (MWD), given the shape, scale, and rate parameters.
The function uses the inverse transformation method to generate random samples from the specified distribution.
}
\details{
The quantile function is the inverse of the cumulative distribution function (CDF) and is often used in simulation studies and reliability analysis.
This function generates random numbers by calculating the quantile function of the Modified Weibull Distribution (MWD), which is particularly useful for data modeling with a variety of real-world applications such as survival analysis, reliability engineering, and risk analysis.
The method used to calculate quantiles is based on the formula for the quantile function:
\deqn{Q(p) = \left( \frac{-log(1 - \frac{log(p(\alpha - 1) + 1)}{log(\alpha)})}{\lambda} \right)^{1/\beta}}
where \eqn{p} is a probability value.
}
\note{
Ensure that the parameters \code{alpha}, \code{beta}, and \code{lambda} are chosen appropriately to avoid numerical issues.
}
\examples{
# Generate 4 random samples from MWD with specified parameters
temp1 <- rqasi(4, 0.2, 0.3, 0.4)

# Generate 40 random samples from MWD with different parameters
temp2 <- rqasi(40, 2, 0.3, 1.4)

}
\references{
Ramzan, Q., Amin, M., & Faisal, M. (2022). Bayesian inference for modified Weibull distribution under simple step‐stress model based on type‐I censoring. \emph{Quality and Reliability Engineering International}, 38(2), 757-779.
}
\seealso{
\code{\link{rweibull}}, \code{\link{runif}}
}
