% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApps.R
\name{run_app}
\alias{run_app}
\title{Application for Simulating Spikes}
\usage{
run_app(app)
}
\arguments{
\item{app}{The type of application you want to open. The options are 'SimulatingSpikes' or 'ViewData'}
}
\description{
Using runExample() produces a GUI for simulating spikes from our model. This allows for the easy simulation of spikes without using R directly, and the generated spike sequences can be locally downloaded.
}
\details{
The application only allows for the implementation of 5 ISI types, namely: Exponential, Gamma, Inverse Gaussian, Log Normal and Weibull. To use alternatively distributions/parameterisations please use or \code{\link{simulate_spikes}}.
}
