% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{run_example}
\alias{run_example}
\title{Application for Simulating Spikes}
\usage{
run_example()
}
\description{
Using runExample() produces a GUI for simulating spikes from our model. This allows for the easy simulation of spikes without using R directly, and the generated spike sequences can be locally downloaded.
}
\details{
The application only allows for the implementation of 5 ISI types, namely: Exponential, Gamma, Inverse Gaussian, Log Normal and Weibull. To use alternatively distributions/parameterisations please use \code{\link{Spikes}} or \code{\link{MultiSpikes}}.
}
