% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Application for Simulating Spikes}
\usage{
runExample()
}
\value{
Shiny app of simulating spikes
}
\description{
Using runExample() produces a GUI for simulating spikes from our model.
}
