% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMC}
\alias{rwMC}
\title{A random walk Metropolis sampler using Rcpp}
\usage{
rwMC(sigma, x0, N)
}
\arguments{
\item{sigma}{the stand deviation of proposal distribution to the random walk Metropolis chains}

\item{x0}{the initial point}

\item{N}{the number of produced values in chain}
}
\value{
a random sample of \code{the standard Laplace distribution}
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution using Rcpp
}
\examples{
\dontrun{
out <- rwMC(2.0,20.0,2000)
plot(out[,1], type="l", main='rwMC(sigma=2)', ylab="x",xlab = "")
}
}
