% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save file as csv}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{the dataset to be saved}

\item{filename}{the name of csv}

\item{row.names}{to suppress row names}

\item{...}{others}
}
\description{
Save file as csv
}
