% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tools.R
\name{shhh_check}
\alias{shhh_check}
\title{Check and Install Packages Quietly}
\usage{
shhh_check(..., quiet = TRUE)
}
\arguments{
\item{...}{Arguments passed to devtools::check or devtools::install.}

\item{quiet}{Logical indicating whether to suppress output (default is TRUE for shhh_check).}
}
\value{
Returns a list of the function results.
}
\description{
Executes the check or install commands quietly, using purrr's quietly wrapper.
}
\examples{
\dontrun{
result <- shhh_check("myPackage")
install_results <- pretty_install("myPackage")
}
}
