% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{show_spikes}
\alias{show_spikes}
\title{Show spikes relative to raw data.}
\usage{
show_spikes(
  data,
  transients = 2000,
  time.step = 2,
  spike.occur = 0.6,
  val.change = NULL,
  reset = spike.occur,
  closeness = 20,
  return.ans = TRUE
)
}
\arguments{
\item{data}{time series data.}

\item{transients}{transient periods where we want to remove spikes.}

\item{time.step}{Vector of time from the time series data, or if step size is regular the time step of the experiment.}

\item{spike.occur}{threshold where we record spikes.}

\item{val.change}{vector of times where the threshold value changes.}

\item{reset}{concentration at which we say a spike is over, default is set to spike.occur.}

\item{closeness}{Value which if two spikes are within this distance they are merged.}

\item{return.ans}{Flag, where if TRUE we return the spikes, else we plot the spikes relative to the raw data.}
}
\value{
Spikes relative to their original position.
}
\description{
Show spikes relative to raw data.
}
