% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_3E.R
\name{sim_3E}
\alias{sim_3E}
\title{Simulates 3 environments with 3-fold and 4-fold cross-validation (CV-O, CV-NO).}
\usage{
sim_3E(omgaG_v0, Y, mu, rho, K, cv, fold, random)
}
\arguments{
\item{omgaG_v0}{A diagonal variance-covariance matrix across environments.}

\item{Y}{A data.frame containing "trait", "environment", and "name".}

\item{mu}{The mean value for each environment.}

\item{rho}{The genetic correlation between environments.}

\item{K}{An additive genetic relationship matrix.}

\item{cv}{Different cross-validation problems, CV-O=1 and CV-NO=2.}

\item{fold}{Different cross-validation methods, such as 3-fold and 4-fold.}

\item{random}{A random seed number.}
}
\description{
Simulates 3 environments with 3-fold and 4-fold cross-validation (CV-O, CV-NO).
}
\examples{

mu <- c(10,15,20)
omgaG_v0 <- diag(c(1,1.5,2))
sim_cvO_3_3E <- sim_3E(omgaG_v0 = omgaG_v0, Y = Yt2.s, mu = mu, rho = 0, K = K.t2, cv = 1,fold = 3,random = 0)
}
