% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_embedding.R
\name{slide_url}
\alias{slide_url}
\title{Construct Slide URL}
\usage{
slide_url(df_url, title, slide = NULL)
}
\arguments{
\item{df_url}{A dataframe containing columns 'link' and 'title'.}

\item{title}{Title of the slide to match in 'df_url'.}

\item{slide}{Optional specific slide number to append to the URL.}
}
\value{
Returns the full URL constructed from the dataframe for the specified title and slide.
}
\description{
Constructs a URL for slides from a dataframe based on the title and optional specific slide number.
}
\examples{
\dontrun{
url <- slide_url(df_slides, "Introduction", 1)
}
}
