% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothCoastData.R
\name{smoothCoastData}
\alias{smoothCoastData}
\alias{buffers}
\alias{world}
\alias{sample_points}
\title{Smoothed coast data objects}
\description{
See \code{vignette("Smooth Coastline Data")} for how these data were created.
}
\details{
\itemize{
\item \code{buffer300} A 300 km buffer list in \code{sp::SpatialPolygons}, \code{sp::SpatialLines}, and
dataframe format in Winkel Tripel projection. A buffer line that is from a coastline
where the islands are removed is also included but this line will go on top of islands
so there will be more upwelling NAs along the coast.
\item \code{buffer20} A 20 km buffer in \code{sp::SpatialPolygons}, \code{sp::SpatialLines}, and
dataframe format in Winkel Tripel projection. A buffer line that is from a coastline
where the islands are removed is also included but this line will go on top of islands
so there will be more upwelling NAs along the coast.
\item \code{world} A world coastline in \code{sp::SpatialLines}, format in Winkel Tripel projection.
\item \code{sample_points} A sample of points from the smooth 20km buffer line where points
are 100km apart.
}
}
\keyword{data}
\keyword{upwelling}
