% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar.R
\name{somar}
\alias{somar}
\title{Soma de dois numeros}
\usage{
somar(x, y)
}
\arguments{
\item{x}{um numero}

\item{y}{um numero}
}
\value{
a soma de \code{x} e \code{y}
}
\description{
\code{somar()} retorna a soma de \code{x} e \code{y}.
}
\examples{
#soma 1+1 = 2
somar (1,1)

}
