% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsowReap.R
\name{Sow}
\alias{Sow}
\title{Asynchroneously return a value}
\usage{
Sow(..., sow_field = "default")
}
\arguments{
\item{...}{value(s) to be returned}

\item{sow_field}{tag to indicate a pool of values into which the value(s) are to be stored, defaults to 'default'}
}
\description{
Asynchroneously return a value
}
\details{
This function stores values into the package environment, so that they can later be
 retrieved by the \code{Reap} function. A ReferenceClass is used to implement the storage logic.
 Nested calls of Reap/Sow are possible as a stack is internally maintained.
}
\examples{

packageDefinition = list(
	name = 'pkg-minimal',
	files = c(),
instFiles = list(),
	description = list(
		title = 'Minimal R-package created with `package`',
		# version to be documented in news section
		#version = '0.1-0',
		author = 'Stefan Böhringer <r-packages@s-boehringer.org>',
		description = 'This appears in the package-documentaion, the markdown of the git-repository and in the package details.',
		depends = c(),
		suggests = c(),
		license = 'LGPL',
		news = "0.1-0	Initial release"
	),
	git = list(
		readme = '## Installation\n```{r}\nlibrary(devtools);\ninstall_github("user/pkg-minimal")\n```\n',
		push = FALSE,
		pushOnNewVersion = FALSE
	)
);


}
\seealso{
package-Reap
}
\author{
Stefan Böhringer, \email{r-packages@s-boehringer.org}
}
\keyword{Reap}
\keyword{Sow}
