% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{spike_merge}
\alias{spike_merge}
\title{Merge spikes}
\usage{
spike_merge(spikes, closeness = 40, data = NULL, time.step = NULL)
}
\arguments{
\item{spikes}{spike sequence}

\item{closeness}{Value which if two spikes are within this distance they are merged.}

\item{data}{time series data the spike sequence is derived from}

\item{time.step}{Vector of time from the time series data, or if step size is regular the time step of the experiment.}
}
\value{
Spike sequence where spikes within closeness have been merged.
}
\description{
Merge spikes
}
