% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_real_data.R
\name{spike_to_ISI}
\alias{spike_to_ISI}
\title{Convert spike times to ISI times}
\usage{
spike_to_ISI(spikes, adjust = FALSE)
}
\arguments{
\item{spikes}{spike times}

\item{adjust}{Flag, where if true we do not include the time to first spike}
}
\value{
ISI times
}
\description{
Convert spike times to ISI times
}
\examples{
spikes <- c(3,9,12,15,17,21)
spike_to_ISI(spikes) #Keep time to first spike
spike_to_ISI(spikes,adjust = TRUE) #Remove time to first spike
}
