% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatingSpikes.R
\name{Spikes}
\alias{Spikes}
\title{Title}
\usage{
Spikes(
  end.time,
  int.fn,
  hyper,
  steps = 1000,
  T.min = NULL,
  ISI.type = "Gamma",
  do.log = T
)
}
\arguments{
\item{end.time}{The length of time we want to simulate spikes}

\item{int.fn}{The intensity function}

\item{hyper}{The ISI parameter}

\item{steps}{The number of steps to split the experiment into}

\item{T.min}{Refractory period (Where pdf = 0)}

\item{ISI.type}{The ISI distribution}

\item{do.log}{Flag for whether we do the calculations on the log scale}
}
\value{
List where $spikes contains the spike sequence and $err gives the max error
}
\description{
Title
}
