% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared.R
\name{squared}
\alias{squared}
\title{Squared a number or a vector}
\usage{
squared(x)
}
\arguments{
\item{x}{a number to be squared}
}
\value{
the square of a number
}
\description{
Squared a number or a vector
}
\examples{
squared(3)
}
