% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianProcess.R
\name{stat_gp}
\alias{stat_gp}
\title{Simulate a Gaussian Process using Spectral decomposition}
\usage{
stat_gp(cov.fn, l, dt.GP = 0.8, steps = 1000)
}
\arguments{
\item{cov.fn}{The covariance function}

\item{l}{the length scale}

\item{dt.GP}{The step size}

\item{steps}{Number of steps required}
}
\value{
A draw from the Gaussian process
}
\description{
Simulate a Gaussian Process using Spectral decomposition
}
