% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split a string into parts (scalar input)}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A string to split. Should be a character vector of length 1.}

\item{split}{A character string containing a regular expression to use for splitting.}
}
\value{
A character vector of the split elements.
}
\description{
A simplified wrapper around \code{strsplit()} for when the input is a single string
and you want a character vector result directly.
}
\examples{
strsplit1("alfa,bravo,charlie,delta", ",")
}
