% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Subtrair de dois numeros}
\usage{
subtrair(x, y)
}
\arguments{
\item{x}{um numero}

\item{y}{um numero}
}
\value{
a subtrair de \code{x} e \code{y}
}
\description{
\code{subtrair()} retorna a subtraicao de \code{x} e \code{y}.
}
\examples{
#subtrair 1-1 = 0
subtrair (1,1)

}
