% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela.R
\name{tabela}
\alias{tabela}
\title{Cria Uma Tabela}
\usage{
tabela(dados)
}
\arguments{
\item{dados}{dataframe}
}
\value{
uma tabela dos \code{dados}
}
\description{
\code{tabela()} retorna uma tabela dos \code{dados}.
}
\examples{
tabela(mtcars)

}
