% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tambah_vektor.R
\name{tambah_vektor}
\alias{tambah_vektor}
\title{Menghitung Penjumlahan 2 vektor}
\usage{
tambah_vektor(u, v)
}
\arguments{
\item{u}{vektor input pertama}

\item{v}{vektor input kedua}
}
\value{
Nilai penjumlahan dari vektor u dan vektor v.
}
\description{
Fungsi ini akan melakukan penjumlahan dari 2 vektor dan menghasilkan hasil grafik.
}
\details{
Fungsi ini untuk menghitung penjumlahan dari 2 vektor yang diinput.
Dimana dilakukan pengecekan terhadap jumlah data dari vektor yang dimasukkan. Kemudian hasil akhir
akan ditampilkan juga dalam bentuk grafik.
}
\examples{
u <- c(1:4)
v <- c(5:8)
tambah_vektor(u,v)
}
