% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Data}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item {y}{a binary outcome}
  \item{x1}{a continuous predictor}
  \item{x2}{a binary predictor}
}
}
\source{
CHL8010 Course Website
}
\usage{
toydata
}
\description{
Data from the course website
}
\references{
Aya Mitani
}
\keyword{datasets}
