% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxDb_to_GTF.R
\name{TxDb_to_GTF}
\alias{TxDb_to_GTF}
\title{Convert exons from the GRangesList to a dataframe compatible with the GTF format.}
\usage{
TxDb_to_GTF()
}
\arguments{
\item{exonsByTx}{A GRangesList}

\item{exonsdf}{A dataframe converted from exonsByTx}

\item{txdbTables}{A list of tables}

\item{initTab}{A merged dataframe of the transcripts and genes tables from txdbTables}

\item{Tab}{A merged datafram of initTab and exondf}

\item{gtf}{A datafram compatible with the GTF format}
}
\value{
gtf
}
\description{
Convert exons from the GRangesList to a dataframe compatible with the GTF format.
}
