% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{check_feasible}
\alias{check_feasible}
\title{Check a step function is feasible.}
\usage{
check_feasible(partition, heights, d.spikes, T.min = NULL)
}
\arguments{
\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{d.spikes}{Data frame containing the spikes.}

\item{T.min}{Refractory period, default set to NULL.}
}
\description{
A function that checks that the partition and heights of a step function is feasible to be used as an intensity function. The function produces an error if the intensity is not feasible.
}
