% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinan_row_reduction.R
\name{determinan_row_reduction}
\alias{determinan_row_reduction}
\title{Determinan dengan Row Reduction}
\usage{
determinan_row_reduction(x)
}
\arguments{
\item{x}{matriks persegi}
}
\value{
Nilai determinan dari matriks
}
\description{
Fungsi ini akan melakukan pencarian nilai determinan dari suatu matriks persegi dengan menggunakan metode row reduction dan disertai dengan langkah pengerjaannya.
}
\details{
Fungsi ini untuk melakukan pencarian nilai determinan dengan menggunakan metode row reduction, Metode
row reduction dilakukan dengan cara transformasi matriks menjadi matriks echelon baris dan kemudian
menerapkan subtitusi mundur untuk mendapatkan nilai determinan.
}
\examples{
x <- matrix(c(0,3,2,1,-6,6,5,9,1),3,3)
determinan_row_reduction(x)
}
