% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_dua_titik.R
\name{distance_dua_titik}
\alias{distance_dua_titik}
\title{Distance antara 2 Titik}
\usage{
distance_dua_titik(point1, point2)
}
\arguments{
\item{point1}{vektor dari titik pertama}

\item{point2}{vektor dari titik kedua}
}
\value{
Nilai jarak dari 2 titik
}
\description{
Fungsi ini akan melakukan pencarian nilai jarak antara 2 titik.
}
\details{
Fungsi ini untuk melakukan pencarian nilai jarak antara 2 titik, dengan melakukan
pengurangan terhadap 2 vektor, kemudian hasil pengurangan dipangkatkan 2, dan dijumlahkan
untuk hasil dari pangkat tersebut. Terakhir untuk mendapatkan nilai jarak dengan
kuadrat dari hasil tersebut.
}
\examples{
point1 <- c(1,2,-2)
point2 <- c(2,4,-4)
distance_dua_titik(point1, point2)
}
