% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_point_plane.R
\name{distance_point_plane}
\alias{distance_point_plane}
\title{Distance antara Titik dan Bidang}
\usage{
distance_point_plane(point, plane)
}
\arguments{
\item{point}{vektor dari titik}

\item{plane}{vektor dari bidang}
}
\value{
Plot jarak dari 1 titik ke 1 bidang
}
\description{
Fungsi ini akan melakukan pencarian nilai jarak antara 1 titik dan 1 bidang,
}
\details{
Fungsi ini untuk melakukan pencarian nilai jarak antara 1 titik dan 1 bidang. DImana titik
berada pada 3 dimensi (x,y,z) dan bidang dengan persamaan ax + by + cz + d = 0 dituliskan
menjadi bentuk vektor (a,b,c,d). Kemudian akan ditampilkan plot untuk jarak antara
titik dengan bidang.

#' Contoh jika ingin mencari jarak pada titik (1,-4,-3) dan persamaan bidnag
2x - 3y + 6z + 1 = 0, maka untuk bidang dituliskan menjadi
(2,-3,6,1), maka untuk menggunakan fungsi ini dengan:

point <- c(1,-4,-3)

plane <- c(2,-3,6,1)

distance_point_plane(point,plane)
}
\examples{
point <- c(1,-4,-3)
plane <- c(2,-3,6,1)
distance_point_plane(point,plane)
}
