% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R
\name{dividir}
\alias{dividir}
\title{Dividir Dois Numeros}
\usage{
dividir(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um nmero}
}
\value{
A divição de \code{x} e \code{y}
}
\description{
\code{dividir()} retorna a divição entre \code{x} e \code{y}.
}
\examples{
# Dividir 4/2 = 2
dividir(4,2)

}
