% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_product.R
\name{dot_product}
\alias{dot_product}
\title{Dot Product Vektor}
\usage{
dot_product(u, v)
}
\arguments{
\item{u}{vektor input pertama}

\item{v}{vektor input pertama}
}
\value{
Plot dari 2 vektor
}
\description{
Fungsi ini akan melakukan pencarian nilai dot product vektor.
}
\details{
Fungsi ini untuk melakukan pencarian nilai dot product vektor.
}
\examples{
u <- c(0,0,1)
v <- c(0,2,2)
dot_product(u,v)
}
