% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{get_X}
\alias{get_X}
\title{get_X()}
\usage{
get_X(
  d.spikes,
  partition = NA,
  heights = NA,
  x = NA,
  end.time = NA,
  T.min = NULL,
  rm.end.time = T,
  int.method = "trapezium"
)
}
\arguments{
\item{d.spikes}{Data frame containing spike sequences}

\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{x}{The intensity function}

\item{end.time}{End of experiment time}

\item{T.min}{Refractory period, default set to NULL.}

\item{rm.end.time}{Flag, where if TRUE we remove the end time of the spike sequences.}

\item{int.method}{Method for integration}
}
\value{
data frame containing the integral of the intensity between all spike times.
}
\description{
Calculates the integral of the intensity function, where the intensity is a piece-wise constant function.
}
