% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatsAroundPoint.R
\name{getStatsAroundPoint}
\alias{getStatsAroundPoint}
\title{Get the mean raster values around a point}
\usage{
getStatsAroundPoint(
  p,
  r,
  d = 100,
  units = "km",
  crs.to.use = crs.wintri,
  fun = "mean"
)
}
\arguments{
\item{p}{the points as a \code{sp::SpatialPoints} object}

\item{r}{the raster as a \code{raster::raster} object}

\item{d}{distance from the coastline}

\item{units}{what units the distance is in}

\item{crs.to.use}{name of the projection. must be one that
the \strong{sp} package recognizes.}
}
\value{
raster value at the points
}
\description{
This is applies \code{raster::extract()} but does some checking on the projection
of the inputs.
}
\keyword{image}
