% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Invers_row_reduction.R
\name{invers_row_reduction}
\alias{invers_row_reduction}
\title{Invers with row reduction}
\usage{
invers_row_reduction(x)
}
\arguments{
\item{x}{squared matrix variable}
}
\description{
'invers' dengan menggunakan metode row reduction secara bertahap
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
x <- matrix(1:4,2,2)
invers_row_reduction(x)
}
