% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shared.R
\name{log_pi_hyper_param}
\alias{log_pi_hyper_param}
\title{Marginal of the ISI parameter}
\usage{
log_pi_hyper_param(
  d.spikes,
  hyper.param,
  partition = NA,
  heights = NA,
  x = NA,
  end.time = NA,
  prior.hyper.param = c(1, 0.01),
  T.min = NULL,
  max.T.min = NA,
  ISI.type = "Gamma"
)
}
\arguments{
\item{d.spikes}{Data frame containing spike sequences.}

\item{hyper.param}{ISI parameter}

\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{x}{Intensity function}

\item{end.time}{End of experiment time}

\item{prior.hyper.param}{Prior of the hyper parameter.}

\item{T.min}{Refractory period, default set to NULL.}

\item{max.T.min}{Maximal allowed T.min values from d.spikes.}

\item{ISI.type}{The ISI distribution.}
}
\value{
marginal of the ISI parameter
}
\description{
Marginal of the ISI parameter
}
