% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWC.R
\name{log_pi_Tmin}
\alias{log_pi_Tmin}
\title{Likelihood of the refratory period}
\usage{
log_pi_Tmin(
  T.min,
  partition,
  heights,
  d.spikes,
  hyper,
  prior.T.min,
  ISI.type = "Gamma"
)
}
\arguments{
\item{T.min}{Refractory period.}

\item{partition}{Partition of time experiment time. A vector of K values where the first entry is 0 and the last is the end of experiment time.}

\item{heights}{Vector of (K-1) heights corresponding to the partition.}

\item{d.spikes}{Data frame containing spike sequences.}

\item{hyper}{ISI parameter.}

\item{prior.T.min}{Prior for the refractory period.}

\item{ISI.type}{The ISI distribution.}
}
\value{
marginal of the refractory period
}
\description{
Likelihood of the refratory period
}
