% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Multiplicar Dois Numeros}
\usage{
multiplicar(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um nmero}
}
\value{
A multiplicação de \code{x} e \code{y}
}
\description{
\code{multiplicar()} retorna a multiplicação entre \code{x} e \code{y}.
}
\examples{
# Multiplicar 2*2 = 4
multiplicar(2,2)

}
