% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividir.R, R/multiplicar.R
\name{multiplicar}
\alias{multiplicar}
\title{Dividir um numero pelo outro}
\usage{
multiplicar(x, y)

multiplicar(x, y)
}
\arguments{
\item{x}{um numero}

\item{y}{um numero}
}
\value{
a divisao de  \code{x} por \code{y}

a multiplica  \code{x} e \code{y}
}
\description{
\code{Dividir()} retorna a divisao de \code{x} e \code{y}.

\code{multiplicar()} retorna a multiplicacao de \code{x} e \code{y}.
}
\examples{
#subtrair 6/*2 = 3
subtrair (6,2)

#subtrair 2*3 = 6
subtrair (2,3)

}
