% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package-package}
\alias{package}
\alias{package-package}
\title{This is package `package`}
\description{
Create packages from R-code directly
}
\details{
This package simplifies package generation by automating the use of `devtools` and `roxygen`. It also makes the development workflow more efficient by allowing ad-hoc development of packages.
}
