% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{package-package}
\alias{package}
\alias{package-package}
\title{This is package `package`}
\description{
Create packages from R-code directly
}
\details{
This package simplifies package generation by automating the use of `devtools` and `roxygen`. It also makes the development workflow more efficient by allowing ad-hoc development of packages. Use `help(package = 'package')` for a tutorial.
This package allows you to create a fully-fledged R-package from a single R-file reducing the added work-load for maintaining a package to a minimum. Depending on the project, collections of files can be used and configuration can be seperated into a stand-alone configuration file, providing full flexibility. It can also handle git interaction.
}
\seealso{
{createPackage()} for starting the main workflow
}
