% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-package.R
\docType{package}
\name{package-package}
\alias{package}
\alias{package-package}
\title{package: raukr2025.package.practice}
\description{
libminer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/XueKang-Qi-1024/package}
  \item Report bugs at \url{https://github.com/XueKang-Qi-1024/package/issues}
}

}
\author{
\strong{Maintainer}: Xuekang Qi \email{xuekang.qi@ki.se}

}
\keyword{internal}
