% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rpackage.R
\docType{data}
\name{packageReadmeTemplate}
\alias{packageReadmeTemplate}
\title{%{name}s}
\format{An object of class \code{character} of length 1.}
\usage{
packageReadmeTemplate
}
\description{
%{title}s
}
\details{
%{description}s

This package allows you to create a fully-fledged R-package from a single R-file reducing the added work-load for maintaining a package to a minimum. Depending on the project, collections of files can be used and configuration can be seperated into a stand-alone cofiguration file, providing full flexibility.
}
\keyword{datasets}
