% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proyeksi_vektor.R
\name{proyeksi_vektor}
\alias{proyeksi_vektor}
\title{Proyeksi Vektor}
\usage{
proyeksi_vektor(a, b)
}
\arguments{
\item{a}{vektor input pertama}

\item{b}{vektor input pertama}
}
\value{
Plot vektor proyeksi dan proyeksi ortogonal.
}
\description{
Fungsi ini akan melakukan pencarian vektor proyeksi dan proyeksi ortogonal.
}
\details{
Fungsi ini untuk melakukan pencarian pencarian vektor proyeksi dan proyeksi ortogonal,
dimana untuk dua vektor yang tidak berhimpit sehingga salah satu vektor bisa
diproyeksikan ke vektor lainnya. Kemudian juga akan ditampilkan ke dalam
bentuk plot grafik dari vektor proyeksi yang didapatkan.
}
\examples{
a <- c(1,3)
b <- c(6,2)
proyeksi_vektor(a,b)
}
