% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_template.r
\name{RFMfunction}
\alias{RFMfunction}
\title{calculateRFM}
\usage{
RFMfunction(
  data,
  weight_recency = 1,
  weight_frequency = 1,
  weight_monetary = 1
)
}
\arguments{
\item{x}{a data object with columns TransDate, Customer, PurchAmount, Cost}

\item{date}{a specification for the time of interest. Must be of format "dd.mm.yyy"}
}
\value{
Returns a data.data containing the recency, frequency and monetary score as well as the weighted final score and the group mebership.
}
\description{
Calculate the a weighted RMF score: recency, frequency, and monetary for every customers.
}
\details{
\code{data} contains the transaction data. The data set must contain a
            column labeled "Customer" that allows unique customer identification
}
