% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtrair.R
\name{subtrair}
\alias{subtrair}
\title{Subtrair Dois Numeros}
\usage{
subtrair(x, y)
}
\arguments{
\item{x}{Um numero}

\item{y}{Um nmero}
}
\value{
A subtração de \code{x} e \code{y}
}
\description{
\code{subtrair()} retorna a subtração entre \code{x} e \code{y}.
}
\examples{
# Subtrair 1-1 = 0
subtrair(1,1)

}
