% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tambah_vektor.R
\name{tambah_vektor}
\alias{tambah_vektor}
\title{Menghitung Penjumlahan 2 vektor}
\usage{
tambah_vektor(u, v)
}
\arguments{
\item{u}{vektor input pertama}

\item{v}{vektor input kedua}
}
\value{
Plot penjumlahan dari vektor u dan vektor v.
}
\description{
Fungsi ini akan melakukan penjumlahan dari 2 vektor dan menghasilkan hasil plot grafik.
}
\details{
Fungsi ini untuk menghitung penjumlahan dari 2 vektor yang diinput.
Dimana dilakukan pengecekan terhadap jumlah elemen dari vektor yang dimasukkan. Kemudian hasil akhir
akan ditampilkan juga dalam bentuk plot grafik.
}
\examples{
u <- c(1,2)
v <- c(5,0)
tambah_vektor(u,v)
}
